refactor: split operator course management services
