+++ src/api/flaskr/service/shifu/admin.py
@@ -1,20 +1,18 @@
 from __future__ import annotations
 
 import json
-import math
 import re
+import sys
 from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
 from decimal import Decimal
 from json import JSONDecodeError
 from typing import Any, Dict, Iterable, Optional, Sequence, Set
 
-from flask import Flask, current_app
-from sqlalchemy import and_, case, false, literal, not_, or_
+from flask import current_app
+from sqlalchemy import and_, case, literal, not_, or_
 from sqlalchemy.orm import defer
 
-from flaskr.common.cache_provider import cache as redis
-from flaskr.common.config import get_config
 from flaskr.common.umami_client import get_course_visit_count_30d
 from flaskr.i18n import _
 from flaskr.dao import db
@@ -58,21 +56,15 @@ from flaskr.service.metering.consts import (
     BILL_USAGE_TYPE_TTS,
 )
 from flaskr.service.metering.models import BillUsageRecord
-from flaskr.service.learn.learn_dtos import ElementType
-from flaskr.service.learn.listen_element_payloads import _deserialize_payload
 from flaskr.service.learn.const import (
     LEARN_STATUS_COMPLETED,
     LEARN_STATUS_RESET,
-    ROLE_STUDENT,
-    ROLE_TEACHER,
 )
 from flaskr.service.learn.models import (
     LearnGeneratedBlock,
     LearnGeneratedElement,
-    LearnLessonFeedback,
     LearnProgressRecord,
 )
-from flaskr.service.common.dtos import PageNationDTO
 from flaskr.service.common.models import (
     raise_error,
     raise_error_with_args,
@@ -80,41 +72,16 @@ from flaskr.service.common.models import (
 )
 from flaskr.service.order.consts import ORDER_STATUS_SUCCESS
 from flaskr.service.order.models import Order
-from flaskr.service.profile.models import Variable
-from flaskr.util import generate_id
 from flaskr.service.shifu.admin_dtos import (
-    AdminOperationCourseListDTO,
-    AdminOperationCourseChapterDetailDTO,
-    AdminOperationCourseDetailBasicInfoDTO,
-    AdminOperationCourseDetailChapterDTO,
-    AdminOperationCourseDetailDTO,
     AdminOperationCourseCreditUsageDetailItemDTO,
-    AdminOperationCourseCreditUsageDetailListDTO,
     AdminOperationCourseCreditUsageItemDTO,
-    AdminOperationCourseCreditUsageListDTO,
-    AdminOperationCourseFollowUpCurrentRecordDTO,
-    AdminOperationCourseFollowUpDetailBasicInfoDTO,
-    AdminOperationCourseFollowUpDetailDTO,
-    AdminOperationCourseFollowUpItemDTO,
-    AdminOperationCourseFollowUpListDTO,
-    AdminOperationCourseFollowUpSummaryDTO,
-    AdminOperationCourseFollowUpTimelineItemDTO,
-    AdminOperationCourseDetailMetricsDTO,
-    AdminOperationCoursePromptDTO,
-    AdminOperationCourseRatingItemDTO,
-    AdminOperationCourseRatingListDTO,
-    AdminOperationCourseRatingSummaryDTO,
-    AdminOperationCourseOverviewDTO,
-    AdminOperationCourseUserDTO,
     AdminOperationCourseSummaryDTO,
     AdminOperationUserCreditLedgerItemDTO,
     AdminOperationUserCreditSummaryDTO,
     AdminOperationUserCourseSummaryDTO,
     AdminOperationUserSummaryDTO,
 )
-from flaskr.service.shifu.course_activity import load_course_activity_map
 from flaskr.service.shifu.consts import (
-    BLOCK_TYPE_MDASK_VALUE,
     BLOCK_TYPE_MDANSWER_VALUE,
     BLOCK_TYPE_MDINTERACTION_VALUE,
     BLOCK_TYPE_MDCONTENT_VALUE,
@@ -123,19 +90,9 @@ from flaskr.service.shifu.consts import (
     UNIT_TYPE_VALUE_NORMAL,
     UNIT_TYPE_VALUE_TRIAL,
 )
-from flaskr.service.shifu.demo_courses import (
-    BUILTIN_DEMO_TITLES,
-    is_builtin_demo_course,
-    load_demo_shifu_bids,
-)
+from flaskr.service.shifu.demo_courses import load_demo_shifu_bids
 from flaskr.service.shifu.shifu_draft_funcs import (
     check_text_with_risk_control,
-    get_latest_shifu_draft,
-)
-from flaskr.service.shifu.shifu_history_manager import (
-    HistoryItem,
-    save_outline_tree_history,
-    save_shifu_history,
 )
 from flaskr.service.shifu.models import (
     AiCourseAuth,
@@ -157,17 +114,9 @@ from flaskr.service.user.models import (
     UserInfo as UserEntity,
     UserToken,
 )
-from flaskr.service.user.repository import (
-    ensure_user_for_identifier,
-    load_user_aggregate_by_identifier,
-    set_user_state,
-    upsert_credential,
-)
 from flaskr.util.timezone import serialize_with_app_timezone
 from flaskr.service.user.utils import (
-    ensure_demo_course_permissions,
     load_existing_demo_shifu_ids,
-    mark_creator_role_if_needed,
     run_creator_granted_post_auth,
 )
 from markdown_flow import MarkdownFlow
@@ -3608,510 +3557,6 @@ def _build_outline_history_tree(
     return _build("")
 
 
-def _copy_course_variable_definitions(
-    *,
-    source_shifu_bid: str,
-    target_shifu_bid: str,
-    creator_user_bid: str,
-    updated_user_bid: str,
-    now: datetime,
-) -> None:
-    variable_definitions = (
-        Variable.query.filter(
-            Variable.shifu_bid == source_shifu_bid,
-            Variable.deleted == 0,
-        )
-        .order_by(Variable.id.asc())
-        .all()
-    )
-    for definition in variable_definitions:
-        db.session.add(
-            Variable(
-                variable_bid=generate_id(current_app),
-                shifu_bid=target_shifu_bid,
-                key=str(definition.key or "").strip(),
-                is_hidden=definition.is_hidden,
-                deleted=0,
-                created_at=now,
-                created_user_bid=creator_user_bid,
-                updated_at=now,
-                updated_user_bid=updated_user_bid,
-            )
-        )
-
-
-def _run_course_copy_draft_risk_check(
-    app: Flask,
-    *,
-    source_draft: DraftShifu,
-    target_shifu_bid: str,
-    operator_user_bid: str,
-    new_course_name: str,
-) -> None:
-    draft_to_check = source_draft.clone()
-    draft_to_check.shifu_bid = target_shifu_bid
-    draft_to_check.title = new_course_name
-    check_content = str(draft_to_check.get_str_to_check() or "").strip()
-    if check_content:
-        check_text_with_risk_control(
-            app,
-            target_shifu_bid,
-            operator_user_bid,
-            check_content,
-        )
-
-
-def _run_course_copy_outline_risk_check(
-    app: Flask,
-    *,
-    source_outline: DraftOutlineItem,
-    target_outline_bid: str,
-    operator_user_bid: str,
-) -> None:
-    outline_to_check = source_outline.clone()
-    outline_to_check.outline_item_bid = target_outline_bid
-    outline_check_content = str(outline_to_check.get_str_to_check() or "").strip()
-    if outline_check_content:
-        check_text_with_risk_control(
-            app,
-            target_outline_bid,
-            operator_user_bid,
-            outline_check_content,
-        )
-
-    markdown_content = str(outline_to_check.content or "").strip()
-    if markdown_content:
-        check_text_with_risk_control(
-            app,
-            target_outline_bid,
-            operator_user_bid,
-            markdown_content,
-        )
-
-
-def _validate_operator_target_contact(contact_type: str, identifier: str) -> str:
-    normalized_contact_type = str(contact_type or "").strip().lower()
-    normalized_identifier = _normalize_identifier(identifier)
-    if normalized_contact_type not in {"phone", "email"}:
-        raise_param_error("contact_type")
-    if (
-        not normalized_identifier
-        or len(normalized_identifier) > OPERATOR_TARGET_CONTACT_MAX_LENGTH
-    ):
-        raise_param_error("contact")
-    if normalized_contact_type == "phone":
-        if not OPERATOR_TARGET_PHONE_PATTERN.match(normalized_identifier):
-            raise_param_error("mobile")
-        return normalized_identifier
-    if not OPERATOR_TARGET_EMAIL_PATTERN.match(normalized_identifier):
-        raise_param_error("email")
-    return normalized_identifier.lower()
-
-
-def _prepare_operator_target_creator(
-    app: Flask,
-    *,
-    contact_type: str,
-    identifier: str,
-    previous_creator_user_bid: str = "",
-    allow_same_user: bool = False,
-) -> Dict[str, Any]:
-    normalized_contact_type = str(contact_type or "").strip().lower()
-    normalized_identifier = _validate_operator_target_contact(
-        normalized_contact_type, identifier
-    )
-
-    lookup_providers = (
-        ["email", "google"] if normalized_contact_type == "email" else ["phone"]
-    )
-
-    existing_aggregate = load_user_aggregate_by_identifier(
-        normalized_identifier,
-        providers=lookup_providers,
-    )
-    created_new_user = False
-    granted_demo_permissions = False
-    if existing_aggregate is None:
-        target_aggregate, created_new_user = ensure_user_for_identifier(
-            app,
-            provider=normalized_contact_type,
-            identifier=normalized_identifier,
-            defaults={
-                "identify": normalized_identifier,
-                "nickname": "",
-                "state": USER_STATE_REGISTERED,
-            },
-        )
-    else:
-        target_aggregate = existing_aggregate
-
-    target_user_bid = str(target_aggregate.user_bid or "").strip()
-    if not target_user_bid:
-        raise_error("server.shifu.transferCreatorTargetNotFound")
-    if (
-        previous_creator_user_bid
-        and not allow_same_user
-        and target_user_bid == previous_creator_user_bid
-    ):
-        raise_error("server.shifu.transferCreatorSameUser")
-
-    should_grant_demo_permissions = created_new_user
-    if (
-        existing_aggregate is not None
-        and existing_aggregate.state == USER_STATE_UNREGISTERED
-    ):
-        set_user_state(target_user_bid, USER_STATE_REGISTERED)
-        should_grant_demo_permissions = True
-
-    upsert_credential(
-        app,
-        user_bid=target_user_bid,
-        provider_name=normalized_contact_type,
-        subject_id=normalized_identifier,
-        subject_format=normalized_contact_type,
-        identifier=normalized_identifier,
-        metadata={},
-        verified=True,
-    )
-
-    if should_grant_demo_permissions:
-        demo_shifu_ids = load_existing_demo_shifu_ids()
-        if demo_shifu_ids:
-            ensure_demo_course_permissions(
-                app,
-                target_user_bid,
-                demo_ids=demo_shifu_ids,
-            )
-            granted_demo_permissions = True
-
-    creator_granted_now = mark_creator_role_if_needed(target_user_bid)
-    return {
-        "target_aggregate": target_aggregate,
-        "target_user_bid": target_user_bid,
-        "normalized_identifier": normalized_identifier,
-        "created_new_user": created_new_user,
-        "granted_demo_permissions": granted_demo_permissions,
-        "creator_granted_now": creator_granted_now,
-    }
-
-
-def _load_recent_learning_active_course_bids(
-    *,
-    since: datetime,
-    shifu_bids: Optional[Sequence[str]] = None,
-) -> Set[str]:
-    query = db.session.query(LearnProgressRecord.shifu_bid).filter(
-        LearnProgressRecord.deleted == 0,
-        LearnProgressRecord.status != LEARN_STATUS_RESET,
-        LearnProgressRecord.created_at >= since,
-    )
-    if shifu_bids is not None:
-        normalized_shifu_bids = [
-            str(shifu_bid or "").strip() for shifu_bid in shifu_bids if shifu_bid
-        ]
-        if not normalized_shifu_bids:
-            return set()
-        query = query.filter(LearnProgressRecord.shifu_bid.in_(normalized_shifu_bids))
-    rows = query.distinct().all()
-    return {
-        str(shifu_bid or "").strip()
-        for (shifu_bid,) in rows
-        if str(shifu_bid or "").strip()
-    }
-
-
-def _load_recent_paid_order_course_bids(
-    *,
-    since: datetime,
-    shifu_bids: Optional[Sequence[str]] = None,
-) -> Set[str]:
-    query = db.session.query(Order.shifu_bid).filter(
-        Order.deleted == 0,
-        Order.status == ORDER_STATUS_SUCCESS,
-        Order.created_at >= since,
-    )
-    if shifu_bids is not None:
-        normalized_shifu_bids = [
-            str(shifu_bid or "").strip() for shifu_bid in shifu_bids if shifu_bid
-        ]
-        if not normalized_shifu_bids:
-            return set()
-        query = query.filter(Order.shifu_bid.in_(normalized_shifu_bids))
-    rows = query.distinct().all()
-    return {
-        str(shifu_bid or "").strip()
-        for (shifu_bid,) in rows
-        if str(shifu_bid or "").strip()
-    }
-
-
-def _clear_shifu_permission_cache(app: Flask, user_id: str, shifu_bid: str) -> None:
-    prefixes = {
-        app.config.get("CACHE_KEY_PREFIX", "") or "",
-        get_config("REDIS_KEY_PREFIX") or "",
-    }
-    for prefix in prefixes:
-        cache_key = f"{prefix}shifu_permission:{user_id}:{shifu_bid}"
-        redis.delete(cache_key)
-
-
-def _clear_shifu_creator_cache(app: Flask, shifu_bid: str) -> None:
-    prefixes = {
-        app.config.get("REDIS_KEY_PREFIX", "") or "",
-        get_config("REDIS_KEY_PREFIX") or "",
-        "ai-shifu",
-    }
-    for prefix in prefixes:
-        cache_key = f"{prefix}:shifu_creator:{shifu_bid}"
-        redis.delete(cache_key)
-
-
-def _update_course_creator_bid(
-    shifu_bid: str,
-    creator_user_bid: str,
-    updated_user_bid: str = "",
-) -> None:
-    draft_values = {DraftShifu.created_user_bid: creator_user_bid}
-    published_values = {PublishedShifu.created_user_bid: creator_user_bid}
-    normalized_updated_user_bid = str(updated_user_bid or "").strip()
-    if normalized_updated_user_bid:
-        updated_at = datetime.now()
-        draft_values[DraftShifu.updated_user_bid] = normalized_updated_user_bid
-        draft_values[DraftShifu.updated_at] = updated_at
-        published_values[PublishedShifu.updated_user_bid] = normalized_updated_user_bid
-        published_values[PublishedShifu.updated_at] = updated_at
-    DraftShifu.query.filter(DraftShifu.shifu_bid == shifu_bid).update(
-        draft_values,
-        synchronize_session=False,
-    )
-    PublishedShifu.query.filter(PublishedShifu.shifu_bid == shifu_bid).update(
-        published_values,
-        synchronize_session=False,
-    )
-
-
-def transfer_operator_course_creator(
-    app: Flask,
-    *,
-    shifu_bid: str,
-    contact_type: str,
-    identifier: str,
-    operator_user_bid: str = "",
-) -> Dict[str, Any]:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        normalized_contact_type = str(contact_type or "").strip().lower()
-        normalized_identifier = _normalize_identifier(identifier)
-        normalized_operator_user_bid = str(operator_user_bid or "").strip()
-
-        latest_course = _load_latest_course_for_transfer(normalized_shifu_bid)
-        if not latest_course:
-            raise_error("server.shifu.shifuNotFound")
-        if not _is_operator_visible_course(latest_course):
-            raise_error("server.shifu.transferCreatorDemoNotAllowed")
-
-        previous_creator_user_bid = str(latest_course.created_user_bid or "").strip()
-        target_creator_result = _prepare_operator_target_creator(
-            app,
-            contact_type=normalized_contact_type,
-            identifier=normalized_identifier,
-            previous_creator_user_bid=previous_creator_user_bid,
-        )
-        target_aggregate = target_creator_result["target_aggregate"]
-        target_user_bid = target_creator_result["target_user_bid"]
-        created_new_user = target_creator_result["created_new_user"]
-        granted_demo_permissions = target_creator_result["granted_demo_permissions"]
-        creator_granted_now = target_creator_result["creator_granted_now"]
-        _update_course_creator_bid(
-            normalized_shifu_bid,
-            target_user_bid,
-            updated_user_bid=normalized_operator_user_bid,
-        )
-        if normalized_operator_user_bid and getattr(latest_course, "id", 0):
-            save_shifu_history(
-                app,
-                normalized_operator_user_bid,
-                normalized_shifu_bid,
-                int(latest_course.id),
-            )
-
-        db.session.commit()
-        if previous_creator_user_bid:
-            _clear_shifu_permission_cache(
-                app, previous_creator_user_bid, normalized_shifu_bid
-            )
-        _clear_shifu_permission_cache(app, target_user_bid, normalized_shifu_bid)
-        _clear_shifu_creator_cache(app, normalized_shifu_bid)
-        if creator_granted_now:
-            run_creator_granted_post_auth(
-                app,
-                user_id=target_user_bid,
-                source="operator_transfer_creator",
-                login_context="admin",
-                created_new_user=created_new_user,
-                language=target_aggregate.user_language,
-            )
-        return {
-            "shifu_bid": normalized_shifu_bid,
-            "previous_creator_user_bid": previous_creator_user_bid,
-            "target_creator_user_bid": target_user_bid,
-            "created_new_user": created_new_user,
-            "granted_demo_permissions": granted_demo_permissions,
-        }
-
-
-def copy_operator_course(
-    app: Flask,
-    *,
-    shifu_bid: str,
-    contact_type: str,
-    identifier: str,
-    operator_user_bid: str,
-    new_course_name: str = "",
-) -> Dict[str, Any]:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        normalized_contact_type = str(contact_type or "").strip().lower()
-        normalized_identifier = _normalize_identifier(identifier)
-        normalized_operator_user_bid = str(operator_user_bid or "").strip()
-        if not normalized_operator_user_bid:
-            raise_param_error("operator_user_bid")
-
-        source_draft = get_latest_shifu_draft(normalized_shifu_bid)
-        if not source_draft:
-            raise_error("server.shifu.copyCourseDraftNotFound")
-        if not _is_operator_visible_course(source_draft):
-            raise_error("server.shifu.copyCourseDemoNotAllowed")
-
-        action_user_bid = normalized_operator_user_bid
-        now = datetime.now()
-        new_shifu_bid = generate_id(app)
-        resolved_new_course_name = _resolve_course_copy_title(
-            source_draft.title,
-            new_course_name,
-        )
-        source_outlines = _load_latest_active_draft_outlines(normalized_shifu_bid)
-        outline_bid_map: Dict[str, str] = {
-            str(item.outline_item_bid or "").strip(): generate_id(app)
-            for item in source_outlines
-        }
-
-        _run_course_copy_draft_risk_check(
-            app,
-            source_draft=source_draft,
-            target_shifu_bid=new_shifu_bid,
-            operator_user_bid=action_user_bid,
-            new_course_name=resolved_new_course_name,
-        )
-        for source_outline in source_outlines:
-            old_outline_bid = str(source_outline.outline_item_bid or "").strip()
-            _run_course_copy_outline_risk_check(
-                app,
-                source_outline=source_outline,
-                target_outline_bid=outline_bid_map[old_outline_bid],
-                operator_user_bid=action_user_bid,
-            )
-
-        target_creator_result = _prepare_operator_target_creator(
-            app,
-            contact_type=normalized_contact_type,
-            identifier=normalized_identifier,
-            previous_creator_user_bid=str(source_draft.created_user_bid or "").strip(),
-            allow_same_user=True,
-        )
-        target_aggregate = target_creator_result["target_aggregate"]
-        target_user_bid = target_creator_result["target_user_bid"]
-        created_new_user = target_creator_result["created_new_user"]
-        granted_demo_permissions = target_creator_result["granted_demo_permissions"]
-        creator_granted_now = target_creator_result["creator_granted_now"]
-
-        new_draft = source_draft.clone()
-        new_draft.shifu_bid = new_shifu_bid
-        new_draft.title = resolved_new_course_name
-        new_draft.created_at = now
-        new_draft.updated_at = now
-        new_draft.created_user_bid = target_user_bid
-        new_draft.updated_user_bid = action_user_bid
-        new_draft.deleted = 0
-        db.session.add(new_draft)
-        db.session.flush()
-
-        source_outline_map = {
-            str(item.outline_item_bid or "").strip(): item for item in source_outlines
-        }
-        copied_outlines: Dict[str, DraftOutlineItem] = {}
-
-        for source_outline in source_outlines:
-            old_outline_bid = str(source_outline.outline_item_bid or "").strip()
-            new_outline_bid = outline_bid_map[old_outline_bid]
-
-            new_outline = source_outline.clone()
-            new_outline.shifu_bid = new_shifu_bid
-            new_outline.outline_item_bid = new_outline_bid
-            new_outline.parent_bid = ""
-            new_outline.prerequisite_item_bids = ""
-            new_outline.created_at = now
-            new_outline.updated_at = now
-            new_outline.created_user_bid = target_user_bid
-            new_outline.updated_user_bid = action_user_bid
-            new_outline.deleted = 0
-            db.session.add(new_outline)
-            db.session.flush()
-            copied_outlines[old_outline_bid] = new_outline
-
-        for old_outline_bid, copied_outline in copied_outlines.items():
-            source_outline = source_outline_map[old_outline_bid]
-            parent_old_bid = str(source_outline.parent_bid or "").strip()
-            if parent_old_bid:
-                copied_outline.parent_bid = outline_bid_map.get(parent_old_bid, "")
-
-            prerequisite_old_bids = [
-                bid.strip()
-                for bid in str(source_outline.prerequisite_item_bids or "").split(",")
-                if bid.strip()
-            ]
-            copied_outline.prerequisite_item_bids = ",".join(
-                outline_bid_map[bid]
-                for bid in prerequisite_old_bids
-                if bid in outline_bid_map
-            )
-
-        save_shifu_history(app, action_user_bid, new_shifu_bid, new_draft.id)
-        outline_tree = _build_outline_history_tree(list(copied_outlines.values()))
-        save_outline_tree_history(
-            app,
-            action_user_bid,
-            new_shifu_bid,
-            outline_tree,
-            new_draft.id,
-        )
-        _copy_course_variable_definitions(
-            source_shifu_bid=normalized_shifu_bid,
-            target_shifu_bid=new_shifu_bid,
-            creator_user_bid=target_user_bid,
-            updated_user_bid=action_user_bid,
-            now=now,
-        )
-
-        db.session.commit()
-        if creator_granted_now:
-            run_creator_granted_post_auth(
-                app,
-                user_id=target_user_bid,
-                source="operator_copy_course",
-                login_context="admin",
-                created_new_user=created_new_user,
-                language=target_aggregate.user_language,
-            )
-
-        return {
-            "source_shifu_bid": normalized_shifu_bid,
-            "new_shifu_bid": new_shifu_bid,
-            "new_course_name": resolved_new_course_name,
-            "target_creator_user_bid": target_user_bid,
-            "created_new_user": created_new_user,
-            "granted_demo_permissions": granted_demo_permissions,
-        }
 
 
 def _merge_courses(
@@ -4713,2882 +4158,172 @@ resolve_operator_user_quick_filter = _resolve_operator_user_quick_filter
 resolve_recent_days_window = _resolve_recent_days_window
 
 
-def _load_operator_course_detail_source(shifu_bid: str):
-    draft, published = _load_latest_course_versions(shifu_bid)
-    visible_draft = draft if draft and _is_operator_visible_course(draft) else None
-    visible_published = (
-        published if published and _is_operator_visible_course(published) else None
-    )
-    if visible_draft is None and visible_published is None:
-        return None
-    return {
-        "course": visible_draft or visible_published,
-        "course_status": (
-            COURSE_STATUS_PUBLISHED if visible_published else COURSE_STATUS_UNPUBLISHED
-        ),
-        "outline_model": DraftOutlineItem if visible_draft else PublishedOutlineItem,
-    }
 
 
-def _load_latest_outline_items(model, shifu_bid: str):
-    latest_subquery = (
-        db.session.query(db.func.max(model.id).label("max_id"))
-        .filter(
-            model.shifu_bid == shifu_bid,
-        )
-        .group_by(model.outline_item_bid)
-        .subquery()
-    )
-    rows = (
-        db.session.query(model)
-        .filter(
-            model.id.in_(db.session.query(latest_subquery.c.max_id)),
-            model.deleted == 0,
-        )
-        .all()
-    )
-
-    def _position_key(item) -> tuple[tuple[int, int | str], ...]:
-        position = str(getattr(item, "position", "") or "").strip()
-        if not position:
-            return ()
-        key_parts: list[tuple[int, int | str]] = []
-        for part in position.split("."):
-            normalized_part = part.strip()
-            if not normalized_part:
-                continue
-            if normalized_part.isdigit():
-                key_parts.append((0, int(normalized_part)))
-            else:
-                key_parts.append((1, normalized_part))
-        return tuple(key_parts)
-
-    return sorted(rows, key=_position_key)
-
-
-def _resolve_learning_permission(item_type: Optional[int]) -> str:
-    if item_type == UNIT_TYPE_VALUE_GUEST:
-        return "guest"
-    if item_type == UNIT_TYPE_VALUE_TRIAL:
-        return "free"
-    if item_type == UNIT_TYPE_VALUE_NORMAL:
-        return "paid"
-    return "unknown"
-
-
-def _resolve_content_status(item) -> str:
-    if str(getattr(item, "content", "") or "").strip():
-        return "has"
-    return "empty"
-
-
-def _resolve_outline_prompt_source(item) -> str:
-    parent_bid = str(getattr(item, "parent_bid", "") or "").strip()
-    if parent_bid:
-        return PROMPT_SOURCE_LESSON
-    return PROMPT_SOURCE_CHAPTER
-
-
-def _resolve_prompt_with_fallback(
-    *,
-    outline_item,
-    outline_item_map: Dict[str, DraftOutlineItem | PublishedOutlineItem],
-    course,
-    field_name: str,
-) -> tuple[str, str]:
-    current_item = outline_item
-    visited_bids: set[str] = set()
-
-    while current_item is not None:
-        prompt_value = str(getattr(current_item, field_name, "") or "").strip()
-        if prompt_value:
-            return prompt_value, _resolve_outline_prompt_source(current_item)
-
-        parent_bid = str(getattr(current_item, "parent_bid", "") or "").strip()
-        if not parent_bid or parent_bid in visited_bids:
-            break
-        visited_bids.add(parent_bid)
-        current_item = outline_item_map.get(parent_bid)
-
-    course_prompt_value = str(getattr(course, field_name, "") or "").strip()
-    if course_prompt_value:
-        return course_prompt_value, PROMPT_SOURCE_COURSE
-
-    return "", ""
-
-
-def _build_chapter_tree(
-    items,
-    user_map: Dict[str, Dict[str, str]],
-    *,
-    follow_up_count_map: Dict[str, int],
-    rating_count_map: Dict[str, int],
-    rating_score_map: Dict[str, str],
-) -> list[AdminOperationCourseDetailChapterDTO]:
-    node_map: Dict[str, AdminOperationCourseDetailChapterDTO] = {}
-    ordered_nodes: list[AdminOperationCourseDetailChapterDTO] = []
-    for item in items:
-        bid = str(item.outline_item_bid or "").strip()
-        if not bid:
-            continue
-        modifier_user_bid = str(getattr(item, "updated_user_bid", "") or "").strip()
-        modifier = user_map.get(modifier_user_bid, {})
-        node = AdminOperationCourseDetailChapterDTO(
-            outline_item_bid=bid,
-            title=item.title or "",
-            parent_bid=item.parent_bid or "",
-            position=item.position or "",
-            node_type="chapter" if not (item.parent_bid or "").strip() else "lesson",
-            learning_permission=_resolve_learning_permission(
-                getattr(item, "type", None)
-            ),
-            is_visible=not bool(getattr(item, "hidden", 0)),
-            content_status=_resolve_content_status(item),
-            follow_up_count=int(follow_up_count_map.get(bid, 0) or 0),
-            rating_score=rating_score_map.get(bid, ""),
-            rating_count=int(rating_count_map.get(bid, 0) or 0),
-            modifier_user_bid=modifier_user_bid,
-            modifier_mobile=modifier.get("mobile", ""),
-            modifier_email=modifier.get("email", ""),
-            modifier_nickname=modifier.get("nickname", ""),
-            updated_at=_format_operator_datetime(item.updated_at),
-            children=[],
-        )
-        node_map[bid] = node
-        ordered_nodes.append(node)
-
-    roots: list[AdminOperationCourseDetailChapterDTO] = []
-    for node in ordered_nodes:
-        parent_bid = node.parent_bid.strip()
-        parent = node_map.get(parent_bid) if parent_bid else None
-        if parent is None:
-            roots.append(node)
-            continue
-        parent.children.append(node)
-
-    def _rollup_learning_stats(
-        node: AdminOperationCourseDetailChapterDTO,
-    ) -> tuple[int, int]:
-        follow_up_count = int(node.follow_up_count or 0)
-        rating_count = int(node.rating_count or 0)
-        for child in node.children:
-            child_follow_up_count, child_rating_count = _rollup_learning_stats(child)
-            follow_up_count += child_follow_up_count
-            rating_count += child_rating_count
-        node.follow_up_count = follow_up_count
-        node.rating_count = rating_count
-        return follow_up_count, rating_count
-
-    for root in roots:
-        _rollup_learning_stats(root)
-    return roots
-
-
-def _load_outline_learning_stats(
-    shifu_bid: str,
-    outline_item_bids: Sequence[str],
-) -> tuple[Dict[str, int], Dict[str, int], Dict[str, str]]:
-    normalized_outline_item_bids = [
-        str(outline_item_bid or "").strip()
-        for outline_item_bid in outline_item_bids
-        if str(outline_item_bid or "").strip()
-    ]
-    if not normalized_outline_item_bids:
-        return {}, {}, {}
-
-    follow_up_rows = (
-        db.session.query(
-            LearnGeneratedBlock.outline_item_bid,
-            db.func.count(LearnGeneratedBlock.id),
-        )
-        .filter(
-            LearnGeneratedBlock.shifu_bid == shifu_bid,
-            LearnGeneratedBlock.deleted == 0,
-            LearnGeneratedBlock.status == 1,
-            LearnGeneratedBlock.type == BLOCK_TYPE_MDASK_VALUE,
-            LearnGeneratedBlock.role == ROLE_STUDENT,
-            LearnGeneratedBlock.outline_item_bid.in_(normalized_outline_item_bids),
-        )
-        .group_by(LearnGeneratedBlock.outline_item_bid)
-        .all()
-    )
-    follow_up_count_map = {
-        str(outline_item_bid or "").strip(): int(count or 0)
-        for outline_item_bid, count in follow_up_rows
-        if str(outline_item_bid or "").strip()
-    }
-
-    rating_rows = (
-        db.session.query(
-            LearnLessonFeedback.outline_item_bid,
-            db.func.count(LearnLessonFeedback.id),
-            db.func.avg(LearnLessonFeedback.score),
-        )
-        .filter(
-            LearnLessonFeedback.shifu_bid == shifu_bid,
-            LearnLessonFeedback.deleted == 0,
-            LearnLessonFeedback.outline_item_bid.in_(normalized_outline_item_bids),
-        )
-        .group_by(LearnLessonFeedback.outline_item_bid)
-        .all()
-    )
-    rating_count_map: Dict[str, int] = {}
-    rating_score_map: Dict[str, str] = {}
-    for outline_item_bid, count, score in rating_rows:
-        normalized_outline_item_bid = str(outline_item_bid or "").strip()
-        if not normalized_outline_item_bid:
-            continue
-        rating_count_map[normalized_outline_item_bid] = int(count or 0)
-        rating_score_map[normalized_outline_item_bid] = _format_average_score(score)
-
-    return follow_up_count_map, rating_count_map, rating_score_map
-
-
-def _load_operator_course_outline_items(
-    shifu_bid: str,
-) -> tuple[dict[str, object], list[DraftOutlineItem | PublishedOutlineItem]]:
-    detail_source = _load_operator_course_detail_source(shifu_bid)
-    if detail_source is None:
-        raise_error("server.shifu.shifuNotFound")
-
-    outline_model = detail_source["outline_model"]
-    outline_items = _load_latest_outline_items(outline_model, shifu_bid)
-
-    return detail_source, outline_items
-
-
-def _resolve_visible_leaf_outline_bids(
-    outline_items: Sequence[DraftOutlineItem | PublishedOutlineItem],
-) -> list[str]:
-    visible_item_bids: Set[str] = set()
-    visible_parent_bids: Set[str] = set()
-    for item in outline_items:
-        if bool(getattr(item, "hidden", 0)):
-            continue
-        outline_item_bid = str(getattr(item, "outline_item_bid", "") or "").strip()
-        parent_bid = str(getattr(item, "parent_bid", "") or "").strip()
-        if not outline_item_bid:
-            continue
-        visible_item_bids.add(outline_item_bid)
-        if parent_bid:
-            visible_parent_bids.add(parent_bid)
-    return sorted(visible_item_bids - visible_parent_bids)
-
-
-def _build_course_outline_context_map(
-    outline_items: Sequence[DraftOutlineItem | PublishedOutlineItem],
-) -> Dict[str, Dict[str, str]]:
-    outline_item_map = {
-        str(getattr(item, "outline_item_bid", "") or "").strip(): item
-        for item in outline_items
-        if str(getattr(item, "outline_item_bid", "") or "").strip()
-    }
-    context_map: Dict[str, Dict[str, str]] = {}
-
-    for outline_item_bid, item in outline_item_map.items():
-        lesson_title = str(getattr(item, "title", "") or "").strip()
-        lesson_outline_item_bid = outline_item_bid
-        chapter_title = lesson_title
-        chapter_outline_item_bid = outline_item_bid
-        current_item = item
-        visited_bids = {outline_item_bid}
-
-        while current_item is not None:
-            parent_bid = str(getattr(current_item, "parent_bid", "") or "").strip()
-            if not parent_bid or parent_bid in visited_bids:
-                break
-            visited_bids.add(parent_bid)
-            parent_item = outline_item_map.get(parent_bid)
-            if parent_item is None:
-                break
-            chapter_title = str(getattr(parent_item, "title", "") or "").strip()
-            chapter_outline_item_bid = parent_bid
-            current_item = parent_item
-
-        context_map[outline_item_bid] = {
-            "chapter_outline_item_bid": chapter_outline_item_bid,
-            "chapter_title": chapter_title,
-            "lesson_outline_item_bid": lesson_outline_item_bid,
-            "lesson_title": lesson_title,
-        }
-
-    return context_map
-
-
-def _build_course_follow_up_base_subquery(shifu_bid: str):
-    return (
-        db.session.query(
-            LearnGeneratedBlock.id.label("id"),
-            LearnGeneratedBlock.generated_block_bid.label("generated_block_bid"),
-            LearnGeneratedBlock.progress_record_bid.label("progress_record_bid"),
-            LearnGeneratedBlock.user_bid.label("user_bid"),
-            LearnGeneratedBlock.outline_item_bid.label("outline_item_bid"),
-            LearnGeneratedBlock.generated_content.label("follow_up_content"),
-            LearnGeneratedBlock.created_at.label("created_at"),
-            db.func.row_number()
-            .over(
-                partition_by=LearnGeneratedBlock.progress_record_bid,
-                order_by=(
-                    LearnGeneratedBlock.created_at.asc(),
-                    LearnGeneratedBlock.id.asc(),
-                ),
-            )
-            .label("turn_index"),
-        )
-        .filter(
-            LearnGeneratedBlock.shifu_bid == shifu_bid,
-            LearnGeneratedBlock.deleted == 0,
-            LearnGeneratedBlock.status == 1,
-            LearnGeneratedBlock.type == BLOCK_TYPE_MDASK_VALUE,
-            LearnGeneratedBlock.role == ROLE_STUDENT,
-        )
-        .subquery()
-    )
-
-
-def _build_follow_up_user_keyword_filter(
-    user_bid_column: Any, keyword: str
-) -> Any | None:
-    normalized = _normalize_identifier(keyword)
-    if not normalized:
-        return None
-
-    credential_match_exists = (
-        db.session.query(AuthCredential.id)
-        .filter(
-            AuthCredential.user_bid == user_bid_column,
-            AuthCredential.deleted == 0,
-            AuthCredential.provider_name.in_(["phone", "email", "google"]),
-            AuthCredential.identifier.ilike(f"%{normalized}%"),
-        )
-        .exists()
-    )
-
-    user_filters = [UserEntity.nickname.ilike(f"%{normalized}%")]
-    if "@" in normalized or normalized.isdigit():
-        user_filters.append(UserEntity.user_identify.ilike(f"%{normalized}%"))
-
-    user_match_exists = (
-        db.session.query(UserEntity.id)
-        .filter(
-            UserEntity.user_bid == user_bid_column,
-            UserEntity.deleted == 0,
-            or_(*user_filters),
-        )
-        .exists()
-    )
-
-    return or_(credential_match_exists, user_match_exists)
-
-
-def _build_credit_usage_user_keyword_filter(
-    user_bid_column: Any, keyword: str
-) -> Any | None:
-    normalized = _normalize_identifier(keyword)
-    if not normalized:
-        return None
-
-    nickname_match_exists = (
-        db.session.query(UserEntity.id)
-        .filter(
-            UserEntity.user_bid == user_bid_column,
-            UserEntity.deleted == 0,
-            UserEntity.nickname.ilike(f"%{normalized}%"),
-        )
-        .exists()
-    )
-
-    if "@" in normalized:
-        credential_identifier_filter = (
-            db.func.lower(AuthCredential.identifier) == normalized
-        )
-        user_identifier_filter = db.func.lower(UserEntity.user_identify) == normalized
-    elif normalized.isdigit():
-        credential_identifier_filter = AuthCredential.identifier == normalized
-        user_identifier_filter = UserEntity.user_identify == normalized
-    else:
-        return nickname_match_exists
-
-    credential_match_exists = (
-        db.session.query(AuthCredential.id)
-        .filter(
-            AuthCredential.user_bid == user_bid_column,
-            AuthCredential.deleted == 0,
-            AuthCredential.provider_name.in_(["phone", "email", "google"]),
-            credential_identifier_filter,
-        )
-        .exists()
-    )
-
-    user_identifier_match_exists = (
-        db.session.query(UserEntity.id)
-        .filter(
-            UserEntity.user_bid == user_bid_column,
-            UserEntity.deleted == 0,
-            user_identifier_filter,
-        )
-        .exists()
-    )
-
-    return or_(
-        nickname_match_exists,
-        credential_match_exists,
-        user_identifier_match_exists,
-    )
-
-
-def _resolve_follow_up_matching_outline_bids(
-    outline_context_map: Dict[str, Dict[str, str]],
-    chapter_keyword: str,
-) -> Optional[Set[str]]:
-    normalized_keyword = str(chapter_keyword or "").strip().lower()
-    if not normalized_keyword:
-        return None
-
-    return {
-        outline_item_bid
-        for outline_item_bid, context in outline_context_map.items()
-        if normalized_keyword
-        in str(context.get("chapter_title", "") or "").strip().lower()
-        or normalized_keyword
-        in str(context.get("lesson_title", "") or "").strip().lower()
-    }
-
-
-def _resolve_follow_up_answer_block(
-    blocks: Sequence[LearnGeneratedBlock],
-    index: int,
-) -> LearnGeneratedBlock | None:
-    ask_position = int(blocks[index].position or 0)
-    for next_block in blocks[index + 1 :]:
-        next_block_type = int(next_block.type or 0)
-        next_block_role = int(next_block.role or 0)
-        if (
-            next_block_type == BLOCK_TYPE_MDASK_VALUE
-            and next_block_role == ROLE_STUDENT
-        ):
-            return None
-        if next_block_type == BLOCK_TYPE_MDANSWER_VALUE:
-            return next_block
-        if (
-            next_block_type == BLOCK_TYPE_MDCONTENT_VALUE
-            and next_block_role == ROLE_TEACHER
-            and int(next_block.position or 0) == ask_position
-        ):
-            return next_block
-    return None
-
-
-def _resolve_follow_up_answer_content(block: LearnGeneratedBlock | None) -> str:
-    if block is None:
-        return ""
-
-    generated_content = str(getattr(block, "generated_content", "") or "").strip()
-    if generated_content:
-        return generated_content
-
-    return str(getattr(block, "block_content_conf", "") or "").strip()
-
-
-def _load_follow_up_groups_for_progress_record(
-    progress_record_bid: str,
-) -> list[dict[str, Any]]:
-    normalized_progress_record_bid = str(progress_record_bid or "").strip()
-    if not normalized_progress_record_bid:
-        return []
-
-    blocks = (
-        LearnGeneratedBlock.query.filter(
-            LearnGeneratedBlock.progress_record_bid == normalized_progress_record_bid,
-            LearnGeneratedBlock.deleted == 0,
-            LearnGeneratedBlock.status == 1,
-            or_(
-                and_(
-                    LearnGeneratedBlock.type == BLOCK_TYPE_MDASK_VALUE,
-                    LearnGeneratedBlock.role == ROLE_STUDENT,
-                ),
-                LearnGeneratedBlock.type == BLOCK_TYPE_MDANSWER_VALUE,
-                and_(
-                    LearnGeneratedBlock.type == BLOCK_TYPE_MDCONTENT_VALUE,
-                    LearnGeneratedBlock.role == ROLE_TEACHER,
-                ),
-            ),
-        )
-        .order_by(LearnGeneratedBlock.created_at.asc(), LearnGeneratedBlock.id.asc())
-        .all()
-    )
-    groups: list[dict[str, Any]] = []
-    for index, block in enumerate(blocks):
-        if (
-            int(block.type or 0) != BLOCK_TYPE_MDASK_VALUE
-            or int(block.role or 0) != ROLE_STUDENT
-        ):
-            continue
-        answer_block = _resolve_follow_up_answer_block(blocks, index)
-        groups.append(
-            {
-                "ask_block": block,
-                "answer_block": answer_block,
-            }
-        )
-    return groups
-
-
-def _resolve_follow_up_source_from_element(
-    *,
-    shifu_bid: str,
-    user_bid: str,
-    progress_record_bid: str,
-    answer_generated_block_bid: str,
-    fallback_position: int,
-    ask_created_at: datetime | None,
-) -> dict[str, Any]:
-    normalized_answer_generated_block_bid = str(
-        answer_generated_block_bid or ""
-    ).strip()
-    normalized_user_bid = str(user_bid or "").strip()
-    normalized_shifu_bid = str(shifu_bid or "").strip()
-    normalized_progress_record_bid = str(progress_record_bid or "").strip()
-    if (
-        not normalized_answer_generated_block_bid
-        or not normalized_user_bid
-        or not normalized_shifu_bid
-        or not normalized_progress_record_bid
-    ):
-        return {}
-
-    follow_up_elements = (
-        LearnGeneratedElement.query.filter(
-            LearnGeneratedElement.generated_block_bid
-            == normalized_answer_generated_block_bid,
-            LearnGeneratedElement.user_bid == normalized_user_bid,
-            LearnGeneratedElement.shifu_bid == normalized_shifu_bid,
-            LearnGeneratedElement.progress_record_bid == normalized_progress_record_bid,
-            LearnGeneratedElement.event_type == "element",
-            LearnGeneratedElement.element_type.in_(
-                [ElementType.ASK.value, ElementType.ANSWER.value]
-            ),
-            LearnGeneratedElement.deleted == 0,
-            LearnGeneratedElement.status == 1,
-        )
-        .order_by(
-            LearnGeneratedElement.sequence_number.asc(),
-            LearnGeneratedElement.run_event_seq.asc(),
-            LearnGeneratedElement.id.asc(),
-        )
-        .all()
-    )
-    if not follow_up_elements:
-        return {}
-
-    anchor_element_bid = ""
-    for row in follow_up_elements:
-        payload = _deserialize_payload(str(getattr(row, "payload", "") or ""))
-        anchor_element_bid = str(
-            getattr(payload, "anchor_element_bid", "") or ""
-        ).strip()
-        if anchor_element_bid:
-            break
-    if not anchor_element_bid:
-        return {}
-
-    anchor_query = LearnGeneratedElement.query.filter(
-        LearnGeneratedElement.shifu_bid == normalized_shifu_bid,
-        LearnGeneratedElement.user_bid == normalized_user_bid,
-        LearnGeneratedElement.progress_record_bid == normalized_progress_record_bid,
-        LearnGeneratedElement.event_type == "element",
-        or_(
-            LearnGeneratedElement.element_bid == anchor_element_bid,
-            LearnGeneratedElement.target_element_bid == anchor_element_bid,
-        ),
-        LearnGeneratedElement.deleted == 0,
-    )
-    if ask_created_at is not None:
-        anchor_query = anchor_query.filter(
-            LearnGeneratedElement.created_at <= ask_created_at
-        )
-    anchor_element = anchor_query.order_by(
-        LearnGeneratedElement.created_at.desc(),
-        LearnGeneratedElement.sequence_number.desc(),
-        LearnGeneratedElement.run_event_seq.desc(),
-        LearnGeneratedElement.id.desc(),
-    ).first()
-    if anchor_element is None:
-        return {
-            "source_output_content": "",
-            "source_output_type": "element",
-            "source_position": int(fallback_position or 0),
-            "source_element_bid": anchor_element_bid,
-            "source_element_type": "",
-        }
-
-    return {
-        "source_output_content": str(getattr(anchor_element, "content_text", "") or ""),
-        "source_output_type": "element",
-        "source_position": int(fallback_position or 0),
-        "source_element_bid": anchor_element_bid,
-        "source_element_type": str(getattr(anchor_element, "element_type", "") or ""),
-    }
-
-
-def _resolve_follow_up_source_from_blocks(
-    ask_block: LearnGeneratedBlock,
-) -> dict[str, Any]:
-    progress_record_bid = str(
-        getattr(ask_block, "progress_record_bid", "") or ""
-    ).strip()
-    if not progress_record_bid:
-        return {}
-
-    position = int(getattr(ask_block, "position", 0) or 0)
-    query = LearnGeneratedBlock.query.filter(
-        LearnGeneratedBlock.progress_record_bid == progress_record_bid,
-        LearnGeneratedBlock.deleted == 0,
-        LearnGeneratedBlock.role == ROLE_TEACHER,
-        LearnGeneratedBlock.position == position,
-        LearnGeneratedBlock.type.in_(
-            [BLOCK_TYPE_MDINTERACTION_VALUE, BLOCK_TYPE_MDCONTENT_VALUE]
-        ),
-    )
-    ask_created_at = getattr(ask_block, "created_at", None)
-    ask_block_id = int(getattr(ask_block, "id", 0) or 0)
-    if ask_created_at is not None and ask_block_id > 0:
-        query = query.filter(
-            or_(
-                LearnGeneratedBlock.created_at < ask_created_at,
-                and_(
-                    LearnGeneratedBlock.created_at == ask_created_at,
-                    LearnGeneratedBlock.id < ask_block_id,
-                ),
-            )
-        )
-    elif ask_block_id > 0:
-        query = query.filter(LearnGeneratedBlock.id < ask_block_id)
-
-    source_block = query.order_by(
-        LearnGeneratedBlock.created_at.desc(),
-        LearnGeneratedBlock.id.desc(),
-    ).first()
-    if source_block is None:
-        return {}
-
-    source_type = (
-        "interaction"
-        if int(getattr(source_block, "type", 0) or 0) == BLOCK_TYPE_MDINTERACTION_VALUE
-        else "content"
-    )
-    if source_type == "interaction":
-        source_content = str(
-            getattr(source_block, "block_content_conf", "") or ""
-        ).strip()
-        if not source_content:
-            source_content = str(getattr(source_block, "generated_content", "") or "")
-    else:
-        source_content = str(
-            getattr(source_block, "generated_content", "") or ""
-        ).strip()
-        if not source_content:
-            source_content = str(getattr(source_block, "block_content_conf", "") or "")
-
-    return {
-        "source_output_content": source_content,
-        "source_output_type": source_type,
-        "source_position": int(getattr(source_block, "position", 0) or 0),
-        "source_element_bid": "",
-        "source_element_type": "",
-    }
-
-
-def _resolve_follow_up_source(
-    *,
-    ask_block: LearnGeneratedBlock,
-    answer_block: LearnGeneratedBlock | None,
-) -> dict[str, Any]:
-    fallback_position = int(getattr(ask_block, "position", 0) or 0)
-    if answer_block is not None:
-        source = _resolve_follow_up_source_from_element(
-            shifu_bid=str(getattr(ask_block, "shifu_bid", "") or ""),
-            user_bid=str(getattr(ask_block, "user_bid", "") or ""),
-            progress_record_bid=str(
-                getattr(ask_block, "progress_record_bid", "") or ""
-            ),
-            answer_generated_block_bid=str(
-                getattr(answer_block, "generated_block_bid", "") or ""
-            ),
-            fallback_position=fallback_position,
-            ask_created_at=getattr(ask_block, "created_at", None),
-        )
-        if source:
-            return source
-
-    source = _resolve_follow_up_source_from_blocks(ask_block)
-    if source:
-        return source
-
-    return {
-        "source_output_content": "",
-        "source_output_type": "",
-        "source_position": fallback_position,
-        "source_element_bid": "",
-        "source_element_type": "",
-    }
-
-
-def _load_course_related_user_bids(
-    shifu_bid: str,
-    *,
-    creator_user_bid: str,
-) -> tuple[Set[str], Set[str]]:
-    order_user_bids = {
-        str(user_bid or "").strip()
-        for (user_bid,) in db.session.query(Order.user_bid)
-        .filter(
-            Order.shifu_bid == shifu_bid,
-            Order.deleted == 0,
-            Order.status == ORDER_STATUS_SUCCESS,
-            Order.user_bid != "",
-        )
-        .all()
-        if str(user_bid or "").strip()
-    }
-    permission_user_bids = {
-        str(user_bid or "").strip()
-        for (user_bid,) in db.session.query(AiCourseAuth.user_id)
-        .filter(
-            AiCourseAuth.course_id == shifu_bid,
-            AiCourseAuth.status == 1,
-            AiCourseAuth.user_id != "",
-        )
-        .all()
-        if str(user_bid or "").strip()
-    }
-    learning_user_bids = {
-        str(user_bid or "").strip()
-        for (user_bid,) in db.session.query(LearnProgressRecord.user_bid)
-        .filter(
-            LearnProgressRecord.shifu_bid == shifu_bid,
-            LearnProgressRecord.deleted == 0,
-            LearnProgressRecord.status != LEARN_STATUS_RESET,
-            LearnProgressRecord.user_bid != "",
-        )
-        .distinct()
-        .all()
-        if str(user_bid or "").strip()
-    }
-
-    learner_user_bids = order_user_bids | permission_user_bids | learning_user_bids
-    related_user_bids = set(learner_user_bids)
-    normalized_creator_user_bid = str(creator_user_bid or "").strip()
-    if normalized_creator_user_bid:
-        related_user_bids.add(normalized_creator_user_bid)
-    return related_user_bids, learner_user_bids
-
-
-def _load_course_user_paid_amount_map(
-    shifu_bid: str,
-    user_bids: Sequence[str],
-) -> Dict[str, Decimal]:
-    normalized_user_bids = [
-        str(user_bid or "").strip()
-        for user_bid in user_bids
-        if str(user_bid or "").strip()
-    ]
-    if not normalized_user_bids:
-        return {}
-
-    counted_order_amount_expr = _build_course_order_amount_expr()
-    rows = (
-        db.session.query(
-            Order.user_bid,
-            db.func.coalesce(db.func.sum(counted_order_amount_expr), 0).label(
-                "total_paid_amount"
-            ),
-        )
-        .filter(
-            Order.shifu_bid == shifu_bid,
-            Order.user_bid.in_(normalized_user_bids),
-            Order.deleted == 0,
-            Order.status == ORDER_STATUS_SUCCESS,
-        )
-        .group_by(Order.user_bid)
-        .all()
-    )
-    return {
-        str(user_bid or "").strip(): Decimal(str(total_paid_amount or 0))
-        for user_bid, total_paid_amount in rows
-        if str(user_bid or "").strip()
-    }
-
-
-def _load_course_user_last_learning_map(
-    shifu_bid: str,
-    user_bids: Sequence[str],
-) -> Dict[str, datetime]:
-    normalized_user_bids = [
-        str(user_bid or "").strip()
-        for user_bid in user_bids
-        if str(user_bid or "").strip()
-    ]
-    if not normalized_user_bids:
-        return {}
-
-    rows = (
-        db.session.query(
-            LearnProgressRecord.user_bid,
-            db.func.max(LearnProgressRecord.updated_at).label("last_learning_at"),
-        )
-        .filter(
-            LearnProgressRecord.shifu_bid == shifu_bid,
-            LearnProgressRecord.user_bid.in_(normalized_user_bids),
-            LearnProgressRecord.deleted == 0,
-            LearnProgressRecord.status != LEARN_STATUS_RESET,
-        )
-        .group_by(LearnProgressRecord.user_bid)
-        .all()
-    )
-    return {
-        str(user_bid or "").strip(): last_learning_at
-        for user_bid, last_learning_at in rows
-        if str(user_bid or "").strip() and last_learning_at
-    }
-
-
-def _load_course_user_joined_at_map(
-    shifu_bid: str,
-    user_bids: Sequence[str],
-    *,
-    creator_user_bid: str,
-    course_created_at: Optional[datetime],
-) -> Dict[str, datetime]:
-    normalized_user_bids = [
-        str(user_bid or "").strip()
-        for user_bid in user_bids
-        if str(user_bid or "").strip()
-    ]
-    if not normalized_user_bids:
-        return {}
-
-    joined_at_map: Dict[str, datetime] = {}
-
-    def _merge_rows(rows: Sequence[tuple[str, Any]]) -> None:
-        for user_bid, joined_at in rows:
-            normalized_user_bid = str(user_bid or "").strip()
-            normalized_joined_at = _coerce_operator_datetime(joined_at)
-            if not normalized_user_bid or normalized_joined_at is None:
-                continue
-            current = joined_at_map.get(normalized_user_bid)
-            if current is None or normalized_joined_at < current:
-                joined_at_map[normalized_user_bid] = normalized_joined_at
-
-    _merge_rows(
-        db.session.query(
-            Order.user_bid,
-            db.func.min(Order.created_at).label("joined_at"),
-        )
-        .filter(
-            Order.shifu_bid == shifu_bid,
-            Order.user_bid.in_(normalized_user_bids),
-            Order.deleted == 0,
-            Order.status == ORDER_STATUS_SUCCESS,
-        )
-        .group_by(Order.user_bid)
-        .all()
-    )
-    _merge_rows(
-        db.session.query(
-            AiCourseAuth.user_id,
-            db.func.min(
-                db.func.coalesce(AiCourseAuth.updated_at, AiCourseAuth.created_at)
-            ).label("joined_at"),
-        )
-        .filter(
-            AiCourseAuth.course_id == shifu_bid,
-            AiCourseAuth.user_id.in_(normalized_user_bids),
-            AiCourseAuth.status == 1,
-        )
-        .group_by(AiCourseAuth.user_id)
-        .all()
-    )
-    _merge_rows(
-        db.session.query(
-            LearnProgressRecord.user_bid,
-            db.func.min(LearnProgressRecord.created_at).label("joined_at"),
-        )
-        .filter(
-            LearnProgressRecord.shifu_bid == shifu_bid,
-            LearnProgressRecord.user_bid.in_(normalized_user_bids),
-            LearnProgressRecord.deleted == 0,
-            LearnProgressRecord.status != LEARN_STATUS_RESET,
-        )
-        .group_by(LearnProgressRecord.user_bid)
-        .all()
-    )
-
-    normalized_creator_user_bid = str(creator_user_bid or "").strip()
-    normalized_course_created_at = _coerce_operator_datetime(course_created_at)
-    if normalized_creator_user_bid and normalized_course_created_at:
-        current = joined_at_map.get(normalized_creator_user_bid)
-        if current is None or normalized_course_created_at < current:
-            joined_at_map[normalized_creator_user_bid] = normalized_course_created_at
-
-    return joined_at_map
-
-
-def _load_course_user_learned_lesson_count_map(
-    shifu_bid: str,
-    user_bids: Sequence[str],
-    leaf_outline_bids: Sequence[str],
-) -> Dict[str, int]:
-    normalized_user_bids = [
-        str(user_bid or "").strip()
-        for user_bid in user_bids
-        if str(user_bid or "").strip()
-    ]
-    normalized_leaf_outline_bids = [
-        str(outline_item_bid or "").strip()
-        for outline_item_bid in leaf_outline_bids
-        if str(outline_item_bid or "").strip()
-    ]
-    if not normalized_user_bids or not normalized_leaf_outline_bids:
-        return {}
-
-    rows = (
-        db.session.query(
-            LearnProgressRecord.user_bid,
-            db.func.count(db.func.distinct(LearnProgressRecord.outline_item_bid)).label(
-                "learned_lesson_count"
-            ),
-        )
-        .filter(
-            LearnProgressRecord.shifu_bid == shifu_bid,
-            LearnProgressRecord.user_bid.in_(normalized_user_bids),
-            LearnProgressRecord.outline_item_bid.in_(normalized_leaf_outline_bids),
-            LearnProgressRecord.deleted == 0,
-            LearnProgressRecord.status != LEARN_STATUS_RESET,
-        )
-        .group_by(LearnProgressRecord.user_bid)
-        .all()
-    )
-    return {
-        str(user_bid or "").strip(): int(learned_lesson_count or 0)
-        for user_bid, learned_lesson_count in rows
-        if str(user_bid or "").strip()
-    }
-
-
-def get_operator_course_detail(
-    app: Flask,
-    *,
-    shifu_bid: str,
-) -> AdminOperationCourseDetailDTO:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        if not normalized_shifu_bid:
-            raise_param_error("shifu_bid is required")
-
-        detail_source, outline_items = _load_operator_course_outline_items(
-            normalized_shifu_bid
-        )
-        course = detail_source["course"]
-        course_status = detail_source["course_status"]
-
-        creator_user_bid = str(course.created_user_bid or "").strip()
-        visit_count_30d = get_course_visit_count_30d(app, normalized_shifu_bid)
-        learner_count = (
-            db.session.query(db.func.count(db.distinct(LearnProgressRecord.user_bid)))
-            .filter(
-                LearnProgressRecord.shifu_bid == normalized_shifu_bid,
-                LearnProgressRecord.deleted == 0,
-                LearnProgressRecord.status != LEARN_STATUS_RESET,
-            )
-            .scalar()
-            or 0
-        )
-        order_amount_expr = _build_course_order_amount_expr()
-        order_summary = (
-            db.session.query(
-                db.func.count(Order.id).label("order_count"),
-                db.func.coalesce(db.func.sum(order_amount_expr), 0).label(
-                    "order_amount"
-                ),
-            )
-            .filter(
-                Order.shifu_bid == normalized_shifu_bid,
-                Order.deleted == 0,
-                Order.status == ORDER_STATUS_SUCCESS,
-            )
-            .first()
-        )
-        follow_up_count = (
-            db.session.query(db.func.count(LearnGeneratedBlock.id))
-            .filter(
-                LearnGeneratedBlock.shifu_bid == normalized_shifu_bid,
-                LearnGeneratedBlock.deleted == 0,
-                LearnGeneratedBlock.status == 1,
-                LearnGeneratedBlock.type == BLOCK_TYPE_MDASK_VALUE,
-                LearnGeneratedBlock.role == ROLE_STUDENT,
-            )
-            .scalar()
-            or 0
-        )
-        rating_score = (
-            db.session.query(db.func.avg(LearnLessonFeedback.score))
-            .filter(
-                LearnLessonFeedback.shifu_bid == normalized_shifu_bid,
-                LearnLessonFeedback.deleted == 0,
-            )
-            .scalar()
-        )
-        detail_user_bids = {
-            user_bid
-            for user_bid in [creator_user_bid]
-            + [
-                str(getattr(item, "updated_user_bid", "") or "")
-                for item in outline_items
-            ]
-            if str(user_bid or "").strip()
-        }
-        detail_user_map = _load_user_map(sorted(detail_user_bids))
-        creator = detail_user_map.get(creator_user_bid, {})
-        outline_learning_stats = _load_outline_learning_stats(
-            normalized_shifu_bid,
-            [
-                str(getattr(item, "outline_item_bid", "") or "")
-                for item in outline_items
-            ],
-        )
-        follow_up_count_map, rating_count_map, rating_score_map = outline_learning_stats
-        visible_leaf_outline_bids = _resolve_visible_leaf_outline_bids(outline_items)
-        credit_metrics = _build_operator_course_credit_metrics(
-            normalized_shifu_bid,
-            visible_leaf_outline_bids,
-        )
-
-        return AdminOperationCourseDetailDTO(
-            basic_info=AdminOperationCourseDetailBasicInfoDTO(
-                shifu_bid=normalized_shifu_bid,
-                course_name=course.title or "",
-                course_status=course_status,
-                creator_user_bid=creator_user_bid,
-                creator_mobile=creator.get("mobile", ""),
-                creator_email=creator.get("email", ""),
-                creator_nickname=creator.get("nickname", ""),
-                created_at=_format_operator_datetime(course.created_at),
-                updated_at=_format_operator_datetime(course.updated_at),
-            ),
-            metrics=AdminOperationCourseDetailMetricsDTO(
-                visit_count_30d=int(visit_count_30d),
-                learner_count=int(learner_count),
-                order_count=int(getattr(order_summary, "order_count", 0) or 0),
-                order_amount=_format_decimal(
-                    Decimal(str(getattr(order_summary, "order_amount", 0) or 0))
-                ),
-                follow_up_count=int(follow_up_count),
-                rating_score=_format_average_score(rating_score),
-                credit_consumed_total=credit_metrics["credit_consumed_total"],
-                credit_usage_count=credit_metrics["credit_usage_count"],
-                credit_user_count=credit_metrics["credit_user_count"],
-                completed_credit_user_count=credit_metrics[
-                    "completed_credit_user_count"
-                ],
-                completed_user_avg_credits=credit_metrics["completed_user_avg_credits"],
-            ),
-            chapters=_build_chapter_tree(
-                outline_items,
-                detail_user_map,
-                follow_up_count_map=follow_up_count_map,
-                rating_count_map=rating_count_map,
-                rating_score_map=rating_score_map,
-            ),
-        )
-
-
-def get_operator_course_prompt(
-    app: Flask,
-    *,
-    shifu_bid: str,
-) -> AdminOperationCoursePromptDTO:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        if not normalized_shifu_bid:
-            raise_param_error("shifu_bid is required")
-
-        detail_source = _load_operator_course_detail_source(normalized_shifu_bid)
-        if detail_source is None:
-            raise_error("server.shifu.shifuNotFound")
-
-        course = detail_source["course"]
-        return AdminOperationCoursePromptDTO(
-            course_prompt=str(getattr(course, "llm_system_prompt", "") or "").strip()
-        )
-
-
-def get_operator_course_users(
-    app: Flask,
-    *,
-    shifu_bid: str,
-    page_index: int,
-    page_size: int,
-    filters: Optional[dict] = None,
-) -> PageNationDTO:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        if not normalized_shifu_bid:
-            raise_param_error("shifu_bid is required")
-
-        safe_page_index = max(int(page_index or 1), 1)
-        safe_page_size = min(
-            max(int(page_size or 20), 1),
-            COURSE_USER_LIST_MAX_PAGE_SIZE,
-        )
-        filters = filters or {}
-
-        detail_source, outline_items = _load_operator_course_outline_items(
-            normalized_shifu_bid
-        )
-        course = detail_source["course"]
-        creator_user_bid = str(course.created_user_bid or "").strip()
-        related_user_bids, learner_user_bids = _load_course_related_user_bids(
-            normalized_shifu_bid,
-            creator_user_bid=creator_user_bid,
-        )
-        if not related_user_bids:
-            return PageNationDTO(safe_page_index, safe_page_size, 0, [])
-
-        ordered_user_bids = sorted(related_user_bids)
-        users = (
-            UserEntity.query.filter(
-                UserEntity.user_bid.in_(ordered_user_bids),
-                UserEntity.deleted == 0,
-            )
-            .order_by(UserEntity.created_at.desc(), UserEntity.id.desc())
-            .all()
-        )
-        if not users:
-            return PageNationDTO(safe_page_index, safe_page_size, 0, [])
-
-        user_bids = [
-            str(user.user_bid or "").strip() for user in users if user.user_bid
-        ]
-        contact_map = _load_course_user_contact_map(user_bids)
-        last_login_map = _load_operator_user_last_login_map(user_bids)
-        paid_amount_map = _load_course_user_paid_amount_map(
-            normalized_shifu_bid, user_bids
-        )
-        last_learning_map = _load_course_user_last_learning_map(
-            normalized_shifu_bid, user_bids
-        )
-        joined_at_map = _load_course_user_joined_at_map(
-            normalized_shifu_bid,
-            user_bids,
-            creator_user_bid=creator_user_bid,
-            course_created_at=getattr(course, "created_at", None),
-        )
-        visible_leaf_outline_bids = _resolve_visible_leaf_outline_bids(outline_items)
-        total_lesson_count = len(visible_leaf_outline_bids)
-        learned_lesson_count_map = _load_course_user_learned_lesson_count_map(
-            normalized_shifu_bid,
-            user_bids,
-            visible_leaf_outline_bids,
-        )
-
-        keyword = str(filters.get("keyword", "") or "").strip().lower()
-        user_role_filter = str(filters.get("user_role", "") or "").strip().lower()
-        learning_status_filter = (
-            str(filters.get("learning_status", "") or "").strip().lower()
-        )
-        payment_status_filter = (
-            str(filters.get("payment_status", "") or "").strip().lower()
-        )
-
-        items_with_sort_keys: list[
-            tuple[
-                tuple[datetime, datetime, datetime, datetime, str],
-                AdminOperationCourseUserDTO,
-            ]
-        ] = []
-        for user in users:
-            user_bid = str(user.user_bid or "").strip()
-            if not user_bid:
-                continue
-            contact = contact_map.get(user_bid, {})
-            learned_lesson_count = int(learned_lesson_count_map.get(user_bid, 0) or 0)
-            learning_status = _resolve_course_user_learning_status(
-                learned_lesson_count=learned_lesson_count,
-                total_lesson_count=total_lesson_count,
-            )
-            total_paid_amount = paid_amount_map.get(user_bid)
-            is_paid = bool(total_paid_amount and total_paid_amount > 0)
-            user_role = _resolve_course_user_role(
-                is_creator=bool(user.is_creator),
-                is_operator=bool(user.is_operator),
-                is_student=user_bid in learner_user_bids,
-            )
-
-            if keyword:
-                haystack = [
-                    user_bid.lower(),
-                    str(contact.get("mobile", "") or "").lower(),
-                    str(contact.get("email", "") or "").lower(),
-                    str(user.nickname or "").lower(),
-                ]
-                if not any(keyword in value for value in haystack if value):
-                    continue
-
-            if (
-                user_role_filter
-                and user_role_filter != "all"
-                and user_role != user_role_filter
-            ):
-                continue
-            if (
-                learning_status_filter
-                and learning_status_filter != "all"
-                and learning_status != learning_status_filter
-            ):
-                continue
-            if payment_status_filter == "paid" and not is_paid:
-                continue
-            if payment_status_filter == "unpaid" and is_paid:
-                continue
-
-            last_learning_at = last_learning_map.get(user_bid)
-            joined_at = joined_at_map.get(user_bid)
-            last_login_at = last_login_map.get(user_bid)
-            dto = AdminOperationCourseUserDTO(
-                user_bid=user_bid,
-                mobile=str(contact.get("mobile", "") or ""),
-                email=str(contact.get("email", "") or ""),
-                nickname=user.nickname or "",
-                user_role=user_role,
-                learned_lesson_count=learned_lesson_count,
-                total_lesson_count=total_lesson_count,
-                learning_status=learning_status,
-                is_paid=is_paid,
-                total_paid_amount=_format_decimal(total_paid_amount),
-                last_learning_at=_format_operator_datetime(last_learning_at),
-                joined_at=_format_operator_datetime(joined_at),
-                last_login_at=_format_operator_datetime(last_login_at),
-            )
-            items_with_sort_keys.append(
-                (
-                    (
-                        last_learning_at or datetime.min,
-                        joined_at or datetime.min,
-                        last_login_at or datetime.min,
-                        getattr(user, "created_at", None) or datetime.min,
-                        user_bid,
-                    ),
-                    dto,
-                )
-            )
-
-        items_with_sort_keys.sort(key=lambda item: item[0], reverse=True)
-        items = [item for _, item in items_with_sort_keys]
-        total = len(items)
-        start = (safe_page_index - 1) * safe_page_size
-        end = start + safe_page_size
-        paged_items = items[start:end]
-        return PageNationDTO(safe_page_index, safe_page_size, total, paged_items)
-
-
-def get_operator_course_credit_usages(
-    app: Flask,
-    *,
-    shifu_bid: str,
-    page_index: int,
-    page_size: int,
-    filters: Optional[dict] = None,
-) -> AdminOperationCourseCreditUsageListDTO:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        if not normalized_shifu_bid:
-            raise_param_error("shifu_bid is required")
-
-        safe_page_index = max(int(page_index or 1), 1)
-        safe_page_size = min(
-            max(int(page_size or 20), 1),
-            COURSE_CREDIT_USAGE_LIST_MAX_PAGE_SIZE,
-        )
-        filters = filters or {}
-
-        _detail_source, outline_items = _load_operator_course_outline_items(
-            normalized_shifu_bid
-        )
-        outline_context_map = _build_course_outline_context_map(outline_items)
-        visible_leaf_outline_bids = _resolve_visible_leaf_outline_bids(outline_items)
-
-        mode_filter = _resolve_course_credit_usage_mode_filter(
-            str(filters.get("mode", "") or "")
-        )
-        scene_filter = _resolve_course_credit_usage_scene_filter(
-            str(filters.get("usage_scene", "") or "")
-        )
-        view = _resolve_course_credit_usage_view(str(filters.get("view", "") or ""))
-
-        if str(filters.get("mode", "") or "").strip() and not mode_filter:
-            raise_param_error("mode")
-        if str(filters.get("usage_scene", "") or "").strip() and not scene_filter:
-            raise_param_error("usage_scene")
-        if str(filters.get("view", "") or "").strip() and not view:
-            raise_param_error("view")
-
-        query = _build_operator_course_credit_usage_base_query(
-            normalized_shifu_bid,
-            outline_item_bids=visible_leaf_outline_bids,
-        )
-        query = _apply_course_credit_usage_filters(query, filters)
-
-        if view == COURSE_CREDIT_USAGE_VIEW_RAW:
-            total = query.count()
-            rows = (
-                query.order_by(
-                    BillUsageRecord.created_at.desc(), BillUsageRecord.id.desc()
-                )
-                .offset((safe_page_index - 1) * safe_page_size)
-                .limit(safe_page_size)
-                .all()
-            )
-            user_map = _load_user_map(
-                sorted(
-                    {
-                        str(getattr(usage_row, "user_bid", "") or "").strip()
-                        for usage_row, _ledger_amount in rows
-                        if str(getattr(usage_row, "user_bid", "") or "").strip()
-                    }
-                )
-            )
-            raw_items: list[AdminOperationCourseCreditUsageItemDTO] = []
-            for usage_row, ledger_amount in rows:
-                model_display = _build_course_credit_usage_model_display(
-                    str(getattr(usage_row, "provider", "") or ""),
-                    str(getattr(usage_row, "model", "") or ""),
-                )
-                raw_items.append(
-                    _build_operator_course_credit_usage_item(
-                        usage_row=usage_row,
-                        ledger_amount=ledger_amount,
-                        user_map=user_map,
-                        outline_context_map=outline_context_map,
-                        group_key=str(getattr(usage_row, "usage_bid", "") or ""),
-                        usage_count=1,
-                        usage_mode=_resolve_course_credit_usage_mode(usage_row),
-                        model_variant_count=1 if model_display else 0,
-                    )
-                )
-            return AdminOperationCourseCreditUsageListDTO(
-                view=COURSE_CREDIT_USAGE_VIEW_RAW,
-                items=raw_items,
-                page=safe_page_index,
-                page_size=safe_page_size,
-                total=total,
-                page_count=math.ceil(total / safe_page_size) if safe_page_size else 0,
-            )
-
-        usage_rows = query.subquery("operator_course_credit_usage_filtered")
-        generation_name_expr = db.func.lower(
-            usage_rows.c.extra["generation_name"].as_string()
-        )
-        usage_mode_expr = case(
-            (
-                usage_rows.c.usage_type == BILL_USAGE_TYPE_TTS,
-                COURSE_CREDIT_USAGE_MODE_LISTEN,
-            ),
-            (
-                and_(
-                    usage_rows.c.usage_type != BILL_USAGE_TYPE_TTS,
-                    or_(
-                        generation_name_expr.contains("/user_follow_ask/"),
-                        generation_name_expr.startswith("lesson_ask/"),
-                        generation_name_expr.startswith("lesson_preview_ask/"),
-                    ),
-                ),
-                COURSE_CREDIT_USAGE_MODE_ASK,
-            ),
-            else_=COURSE_CREDIT_USAGE_MODE_LEARN,
-        ).label("usage_mode")
-        usage_scene_expr = case(
-            (
-                usage_rows.c.usage_scene == BILL_USAGE_SCENE_DEBUG,
-                COURSE_CREDIT_USAGE_SCENE_DEBUG,
-            ),
-            (
-                usage_rows.c.usage_scene == BILL_USAGE_SCENE_PREVIEW,
-                COURSE_CREDIT_USAGE_SCENE_PREVIEW,
-            ),
-            (
-                usage_rows.c.usage_scene == BILL_USAGE_SCENE_PROD,
-                COURSE_CREDIT_USAGE_SCENE_LEARNING,
-            ),
-            else_="",
-        ).label("usage_scene")
-        group_key_expr = db.func.concat(
-            usage_rows.c.user_bid,
-            literal(":"),
-            usage_rows.c.outline_item_bid,
-            literal(":"),
-            usage_scene_expr,
-            literal(":"),
-            usage_mode_expr,
-        ).label("group_key")
-
-        grouped_query = (
-            db.session.query(
-                group_key_expr,
-                usage_rows.c.user_bid.label("user_bid"),
-                usage_rows.c.outline_item_bid.label("outline_item_bid"),
-                usage_scene_expr,
-                usage_mode_expr,
-                db.func.count(db.func.distinct(usage_rows.c.usage_bid)).label(
-                    "usage_count"
-                ),
-                db.func.count(
-                    db.func.distinct(
-                        db.func.nullif(
-                            db.func.concat(
-                                db.func.coalesce(usage_rows.c.provider, ""),
-                                literal("/"),
-                                db.func.coalesce(usage_rows.c.model, ""),
-                            ),
-                            "/",
-                        )
-                    )
-                ).label("model_variant_count"),
-                db.func.coalesce(
-                    db.func.sum(db.func.abs(usage_rows.c.ledger_amount)), 0
-                ).label("consumed_credits"),
-                db.func.max(usage_rows.c.created_at).label("created_at"),
-            )
-            .select_from(usage_rows)
-            .group_by(
-                group_key_expr,
-                usage_rows.c.user_bid,
-                usage_rows.c.outline_item_bid,
-                usage_scene_expr,
-                usage_mode_expr,
-            )
-        )
-        latest_usage_query = db.session.query(
-            group_key_expr,
-            usage_rows.c.usage_bid.label("usage_bid"),
-            usage_rows.c.progress_record_bid.label("progress_record_bid"),
-            usage_rows.c.generated_block_bid.label("generated_block_bid"),
-            usage_rows.c.provider.label("provider"),
-            usage_rows.c.model.label("model"),
-            db.func.row_number()
-            .over(
-                partition_by=[
-                    usage_rows.c.user_bid,
-                    usage_rows.c.outline_item_bid,
-                    usage_scene_expr,
-                    usage_mode_expr,
-                ],
-                order_by=[usage_rows.c.created_at.desc(), usage_rows.c.id.desc()],
-            )
-            .label("row_number"),
-        ).select_from(usage_rows)
-
-        grouped_subquery = grouped_query.subquery("operator_course_credit_usage_groups")
-        latest_usage_subquery = latest_usage_query.subquery(
-            "operator_course_credit_usage_latest_rows"
-        )
-        total = (
-            db.session.query(db.func.count()).select_from(grouped_subquery).scalar()
-            or 0
-        )
-        grouped_rows = (
-            db.session.query(
-                grouped_subquery.c.group_key,
-                latest_usage_subquery.c.usage_bid,
-                latest_usage_subquery.c.progress_record_bid,
-                latest_usage_subquery.c.generated_block_bid,
-                grouped_subquery.c.user_bid,
-                grouped_subquery.c.outline_item_bid,
-                grouped_subquery.c.usage_scene,
-                grouped_subquery.c.usage_mode,
-                latest_usage_subquery.c.provider,
-                latest_usage_subquery.c.model,
-                grouped_subquery.c.usage_count,
-                grouped_subquery.c.model_variant_count,
-                grouped_subquery.c.consumed_credits,
-                grouped_subquery.c.created_at,
-            )
-            .join(
-                latest_usage_subquery,
-                and_(
-                    latest_usage_subquery.c.group_key == grouped_subquery.c.group_key,
-                    latest_usage_subquery.c.row_number == 1,
-                ),
-            )
-            .order_by(
-                grouped_subquery.c.created_at.desc(), grouped_subquery.c.group_key.asc()
-            )
-            .offset((safe_page_index - 1) * safe_page_size)
-            .limit(safe_page_size)
-            .all()
-        )
-        user_map = _load_user_map(
-            sorted(
-                {
-                    str(getattr(row, "user_bid", "") or "").strip()
-                    for row in grouped_rows
-                    if str(getattr(row, "user_bid", "") or "").strip()
-                }
-            )
-        )
-
-        grouped_items: list[AdminOperationCourseCreditUsageItemDTO] = []
-        for row in grouped_rows:
-            context = outline_context_map.get(
-                str(getattr(row, "outline_item_bid", "") or "").strip(),
-                {
-                    "chapter_outline_item_bid": "",
-                    "chapter_title": "",
-                    "lesson_outline_item_bid": str(
-                        getattr(row, "outline_item_bid", "") or ""
-                    ),
-                    "lesson_title": "",
-                },
-            )
-            user_bid = str(getattr(row, "user_bid", "") or "").strip()
-            user = user_map.get(user_bid, {})
-            grouped_items.append(
-                AdminOperationCourseCreditUsageItemDTO(
-                    group_key=str(getattr(row, "group_key", "") or ""),
-                    usage_bid=str(getattr(row, "usage_bid", "") or ""),
-                    progress_record_bid=str(
-                        getattr(row, "progress_record_bid", "") or ""
-                    ),
-                    generated_block_bid=str(
-                        getattr(row, "generated_block_bid", "") or ""
-                    ),
-                    user_bid=user_bid,
-                    mobile=str(user.get("mobile", "") or ""),
-                    email=str(user.get("email", "") or ""),
-                    nickname=str(user.get("nickname", "") or ""),
-                    chapter_outline_item_bid=str(
-                        context.get("chapter_outline_item_bid", "") or ""
-                    ),
-                    chapter_title=str(context.get("chapter_title", "") or ""),
-                    lesson_outline_item_bid=str(
-                        context.get("lesson_outline_item_bid", "") or ""
-                    ),
-                    lesson_title=str(context.get("lesson_title", "") or ""),
-                    usage_scene=str(getattr(row, "usage_scene", "") or ""),
-                    usage_mode=str(getattr(row, "usage_mode", "") or ""),
-                    provider=str(getattr(row, "provider", "") or ""),
-                    model=str(getattr(row, "model", "") or ""),
-                    usage_count=int(getattr(row, "usage_count", 0) or 0),
-                    model_variant_count=int(
-                        getattr(row, "model_variant_count", 0) or 0
-                    ),
-                    consumed_credits=credit_decimal_to_number(
-                        Decimal(str(getattr(row, "consumed_credits", 0) or 0))
-                    ),
-                    created_at=_format_operator_datetime(
-                        getattr(row, "created_at", None)
-                    ),
-                )
-            )
-
-        return AdminOperationCourseCreditUsageListDTO(
-            view=COURSE_CREDIT_USAGE_VIEW_GROUPED,
-            items=grouped_items,
-            page=safe_page_index,
-            page_size=safe_page_size,
-            total=int(total or 0),
-            page_count=math.ceil(int(total or 0) / safe_page_size)
-            if safe_page_size
-            else 0,
-        )
-
-
-def get_operator_course_credit_usage_details(
-    app: Flask,
-    *,
-    shifu_bid: str,
-    page_index: int,
-    page_size: int,
-    filters: Optional[dict] = None,
-) -> AdminOperationCourseCreditUsageDetailListDTO:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        if not normalized_shifu_bid:
-            raise_param_error("shifu_bid is required")
-
-        safe_page_index = max(int(page_index or 1), 1)
-        safe_page_size = min(max(int(page_size or 10), 1), 50)
-        filters = filters or {}
-        user_bid = str(filters.get("user_bid", "") or "").strip()
-        outline_item_bid = str(filters.get("outline_item_bid", "") or "").strip()
-        mode_filter = _resolve_course_credit_usage_mode_filter(
-            str(filters.get("mode", "") or "")
-        )
-        scene_filter = _resolve_course_credit_usage_scene_filter(
-            str(filters.get("usage_scene", "") or "")
-        )
-        if not user_bid:
-            raise_param_error("user_bid")
-        if not outline_item_bid:
-            raise_param_error("outline_item_bid")
-        if str(filters.get("mode", "") or "").strip() and not mode_filter:
-            raise_param_error("mode")
-        if str(filters.get("usage_scene", "") or "").strip() and not scene_filter:
-            raise_param_error("usage_scene")
-
-        _detail_source, outline_items = _load_operator_course_outline_items(
-            normalized_shifu_bid
-        )
-        visible_leaf_outline_bids = _resolve_visible_leaf_outline_bids(outline_items)
-        query = _build_operator_course_credit_usage_base_query(
-            normalized_shifu_bid,
-            outline_item_bids=visible_leaf_outline_bids,
-        )
-        query = query.filter(
-            BillUsageRecord.user_bid == user_bid,
-            BillUsageRecord.outline_item_bid == outline_item_bid,
-        )
-        query = _apply_course_credit_usage_filters(
-            query, {"mode": mode_filter, "usage_scene": scene_filter}
-        )
-
-        total = query.count()
-        rows = (
-            query.order_by(BillUsageRecord.created_at.desc(), BillUsageRecord.id.desc())
-            .offset((safe_page_index - 1) * safe_page_size)
-            .limit(safe_page_size)
-            .all()
-        )
-        output_summary_map = _load_course_credit_usage_output_summary_map(
-            [usage_row for usage_row, _ledger_amount in rows]
-        )
-        return AdminOperationCourseCreditUsageDetailListDTO(
-            items=[
-                _build_operator_course_credit_usage_detail_item(
-                    usage_row=usage_row,
-                    ledger_amount=ledger_amount,
-                    output_summary=output_summary_map.get(
-                        str(getattr(usage_row, "usage_bid", "") or "").strip(),
-                        "",
-                    ),
-                )
-                for usage_row, ledger_amount in rows
-            ],
-            page=safe_page_index,
-            page_size=safe_page_size,
-            total=total,
-            page_count=math.ceil(total / safe_page_size) if safe_page_size else 0,
-        )
-
-
-def get_operator_course_follow_ups(
-    app: Flask,
-    *,
-    shifu_bid: str,
-    page_index: int,
-    page_size: int,
-    filters: Optional[dict] = None,
-    include_summary: bool = True,
-) -> AdminOperationCourseFollowUpListDTO:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        if not normalized_shifu_bid:
-            raise_param_error("shifu_bid is required")
-
-        safe_page_index = max(int(page_index or 1), 1)
-        safe_page_size = min(
-            max(int(page_size or 20), 1),
-            COURSE_FOLLOW_UP_LIST_MAX_PAGE_SIZE,
-        )
-        filters = filters or {}
-
-        _detail_source, outline_items = _load_operator_course_outline_items(
-            normalized_shifu_bid
-        )
-        outline_context_map = _build_course_outline_context_map(outline_items)
-
-        keyword = str(filters.get("keyword", "") or "").strip()
-        chapter_keyword = str(filters.get("chapter_keyword", "") or "").strip().lower()
-        start_time = filters.get("start_time")
-        end_time = filters.get("end_time")
-        follow_up_base = _build_course_follow_up_base_subquery(normalized_shifu_bid)
-        user_keyword_filter = _build_follow_up_user_keyword_filter(
-            follow_up_base.c.user_bid,
-            keyword,
-        )
-        matching_outline_item_bids = _resolve_follow_up_matching_outline_bids(
-            outline_context_map,
-            chapter_keyword,
-        )
-
-        if chapter_keyword and not matching_outline_item_bids:
-            return AdminOperationCourseFollowUpListDTO(
-                summary=AdminOperationCourseFollowUpSummaryDTO(
-                    follow_up_count=0,
-                    user_count=0,
-                    lesson_count=0,
-                    latest_follow_up_at="",
-                ),
-                items=[],
-                page=safe_page_index,
-                page_size=safe_page_size,
-                total=0,
-                page_count=0,
-            )
-
-        filtered_query = db.session.query(follow_up_base)
-        if user_keyword_filter is not None:
-            filtered_query = filtered_query.filter(user_keyword_filter)
-        if matching_outline_item_bids is not None:
-            filtered_query = filtered_query.filter(
-                follow_up_base.c.outline_item_bid.in_(
-                    sorted(matching_outline_item_bids)
-                )
-            )
-        if start_time:
-            filtered_query = filtered_query.filter(
-                follow_up_base.c.created_at >= start_time
-            )
-        if end_time:
-            filtered_query = filtered_query.filter(
-                follow_up_base.c.created_at <= end_time
-            )
-
-        filtered_follow_ups = filtered_query.subquery()
-        if include_summary:
-            summary_row = db.session.query(
-                db.func.count(filtered_follow_ups.c.id).label("follow_up_count"),
-                db.func.count(
-                    db.func.distinct(db.func.nullif(filtered_follow_ups.c.user_bid, ""))
-                ).label("user_count"),
-                db.func.count(
-                    db.func.distinct(
-                        db.func.nullif(filtered_follow_ups.c.outline_item_bid, "")
-                    )
-                ).label("lesson_count"),
-                db.func.max(filtered_follow_ups.c.created_at).label(
-                    "latest_follow_up_at"
-                ),
-            ).one()
-            total = int(getattr(summary_row, "follow_up_count", 0) or 0)
-        else:
-            summary_row = None
-            total = int(
-                db.session.query(db.func.count(filtered_follow_ups.c.id)).scalar() or 0
-            )
-
-        if total == 0:
-            return AdminOperationCourseFollowUpListDTO(
-                summary=AdminOperationCourseFollowUpSummaryDTO(),
-                items=[],
-                page=safe_page_index,
-                page_size=safe_page_size,
-                total=0,
-                page_count=0,
-            )
-
-        start = (safe_page_index - 1) * safe_page_size
-        paged_rows = (
-            db.session.query(filtered_follow_ups)
-            .order_by(
-                filtered_follow_ups.c.created_at.desc(),
-                filtered_follow_ups.c.id.desc(),
-            )
-            .offset(start)
-            .limit(safe_page_size)
-            .all()
-        )
-        user_map = _load_user_map(
-            sorted(
-                {
-                    str(getattr(row, "user_bid", "") or "").strip()
-                    for row in paged_rows
-                    if str(getattr(row, "user_bid", "") or "").strip()
-                }
-            )
-        )
-
-        items: list[AdminOperationCourseFollowUpItemDTO] = []
-        for row in paged_rows:
-            generated_block_bid = str(
-                getattr(row, "generated_block_bid", "") or ""
-            ).strip()
-            outline_item_bid = str(getattr(row, "outline_item_bid", "") or "").strip()
-            user_bid = str(getattr(row, "user_bid", "") or "").strip()
-            created_at = getattr(row, "created_at", None)
-            context = outline_context_map.get(
-                outline_item_bid,
-                {
-                    "chapter_outline_item_bid": "",
-                    "chapter_title": "",
-                    "lesson_outline_item_bid": outline_item_bid,
-                    "lesson_title": "",
-                },
-            )
-            user = user_map.get(user_bid, {})
-            items.append(
-                AdminOperationCourseFollowUpItemDTO(
-                    generated_block_bid=generated_block_bid,
-                    progress_record_bid=str(
-                        getattr(row, "progress_record_bid", "") or ""
-                    ),
-                    user_bid=user_bid,
-                    mobile=str(user.get("mobile", "") or ""),
-                    email=str(user.get("email", "") or ""),
-                    nickname=str(user.get("nickname", "") or ""),
-                    chapter_outline_item_bid=str(
-                        context.get("chapter_outline_item_bid", "") or ""
-                    ),
-                    chapter_title=str(context.get("chapter_title", "") or ""),
-                    lesson_outline_item_bid=str(
-                        context.get("lesson_outline_item_bid", "") or ""
-                    ),
-                    lesson_title=str(context.get("lesson_title", "") or ""),
-                    follow_up_content=str(getattr(row, "follow_up_content", "") or ""),
-                    turn_index=int(getattr(row, "turn_index", 0) or 0),
-                    created_at=_format_operator_datetime(created_at),
-                )
-            )
-        if include_summary:
-            summary = AdminOperationCourseFollowUpSummaryDTO(
-                follow_up_count=total,
-                user_count=int(getattr(summary_row, "user_count", 0) or 0),
-                lesson_count=int(getattr(summary_row, "lesson_count", 0) or 0),
-                latest_follow_up_at=_format_operator_datetime(
-                    getattr(summary_row, "latest_follow_up_at", None)
-                ),
-            )
-        else:
-            summary = AdminOperationCourseFollowUpSummaryDTO(follow_up_count=total)
-        return AdminOperationCourseFollowUpListDTO(
-            summary=summary,
-            items=items,
-            page=safe_page_index,
-            page_size=safe_page_size,
-            total=total,
-            page_count=math.ceil(total / safe_page_size) if safe_page_size else 0,
-        )
-
-
-def get_operator_course_ratings(
-    app: Flask,
-    *,
-    shifu_bid: str,
-    page_index: int,
-    page_size: int,
-    filters: Optional[dict] = None,
-    include_summary: bool = True,
-) -> AdminOperationCourseRatingListDTO:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        if not normalized_shifu_bid:
-            raise_param_error("shifu_bid is required")
-
-        safe_page_index = max(int(page_index or 1), 1)
-        safe_page_size = min(
-            max(int(page_size or 20), 1),
-            COURSE_RATING_LIST_MAX_PAGE_SIZE,
-        )
-        filters = filters or {}
-
-        _detail_source, outline_items = _load_operator_course_outline_items(
-            normalized_shifu_bid
-        )
-        outline_context_map = _build_course_outline_context_map(outline_items)
-        keyword = _normalize_identifier(str(filters.get("keyword", "") or "")).lower()
-        chapter_keyword = str(filters.get("chapter_keyword", "") or "").strip().lower()
-        score_filter = str(filters.get("score", "") or "").strip()
-        mode_filter = _resolve_course_rating_mode(str(filters.get("mode", "") or ""))
-        has_comment_filter = str(filters.get("has_comment", "") or "").strip().lower()
-        sort_by = _resolve_course_rating_sort_by(str(filters.get("sort_by", "") or ""))
-        start_time = filters.get("start_time")
-        end_time = filters.get("end_time")
-
-        normalized_score_filter: Optional[int] = None
-        if score_filter:
-            if score_filter not in {"1", "2", "3", "4", "5"}:
-                raise_param_error("score")
-            normalized_score_filter = int(score_filter)
-        if str(filters.get("mode", "") or "").strip() and not mode_filter:
-            raise_param_error("mode")
-        if has_comment_filter and has_comment_filter != "true":
-            raise_param_error("has_comment")
-        if str(filters.get("sort_by", "") or "").strip() and not sort_by:
-            raise_param_error("sort_by")
-
-        rated_at_expression = db.func.coalesce(
-            LearnLessonFeedback.updated_at,
-            LearnLessonFeedback.created_at,
-        )
-        base_filters = [
-            LearnLessonFeedback.shifu_bid == normalized_shifu_bid,
-            LearnLessonFeedback.deleted == 0,
-        ]
-
-        user_keyword_filter = _build_follow_up_user_keyword_filter(
-            LearnLessonFeedback.user_bid,
-            keyword,
-        )
-        if user_keyword_filter is not None:
-            base_filters.append(user_keyword_filter)
-
-        matching_outline_item_bids = _resolve_follow_up_matching_outline_bids(
-            outline_context_map,
-            chapter_keyword,
-        )
-        if matching_outline_item_bids is not None:
-            if not matching_outline_item_bids:
-                return AdminOperationCourseRatingListDTO(
-                    summary=AdminOperationCourseRatingSummaryDTO(),
-                    items=[],
-                    page=safe_page_index,
-                    page_size=safe_page_size,
-                    total=0,
-                    page_count=0,
-                )
-            base_filters.append(
-                LearnLessonFeedback.outline_item_bid.in_(
-                    sorted(matching_outline_item_bids)
-                )
-            )
-
-        if normalized_score_filter is not None:
-            base_filters.append(LearnLessonFeedback.score == normalized_score_filter)
-        if mode_filter:
-            base_filters.append(LearnLessonFeedback.mode == mode_filter)
-        if has_comment_filter == "true":
-            base_filters.append(
-                db.func.trim(db.func.coalesce(LearnLessonFeedback.comment, "")) != ""
-            )
-        if start_time:
-            base_filters.append(rated_at_expression >= start_time)
-        if end_time:
-            base_filters.append(rated_at_expression <= end_time)
-
-        summary_source = (
-            db.session.query(
-                LearnLessonFeedback.id.label("id"),
-                LearnLessonFeedback.score.label("score"),
-                LearnLessonFeedback.user_bid.label("user_bid"),
-                rated_at_expression.label("rated_at"),
-            )
-            .filter(*base_filters)
-            .subquery()
-        )
-        if include_summary:
-            summary_row = db.session.query(
-                db.func.avg(summary_source.c.score).label("average_score"),
-                db.func.count(summary_source.c.id).label("rating_count"),
-                db.func.count(
-                    db.func.distinct(db.func.nullif(summary_source.c.user_bid, ""))
-                ).label("user_count"),
-                db.func.max(summary_source.c.rated_at).label("latest_rated_at"),
-            ).one()
-            total = int(getattr(summary_row, "rating_count", 0) or 0)
-        else:
-            summary_row = None
-            total = int(
-                db.session.query(db.func.count(summary_source.c.id)).scalar() or 0
-            )
-
-        if total == 0:
-            return AdminOperationCourseRatingListDTO(
-                summary=AdminOperationCourseRatingSummaryDTO(),
-                items=[],
-                page=safe_page_index,
-                page_size=safe_page_size,
-                total=0,
-                page_count=0,
-            )
-
-        start = (safe_page_index - 1) * safe_page_size
-        page_query = db.session.query(
-            LearnLessonFeedback.id.label("id"),
-            LearnLessonFeedback.lesson_feedback_bid.label("lesson_feedback_bid"),
-            LearnLessonFeedback.progress_record_bid.label("progress_record_bid"),
-            LearnLessonFeedback.user_bid.label("user_bid"),
-            LearnLessonFeedback.outline_item_bid.label("outline_item_bid"),
-            LearnLessonFeedback.score.label("score"),
-            LearnLessonFeedback.comment.label("comment"),
-            LearnLessonFeedback.mode.label("mode"),
-            rated_at_expression.label("rated_at"),
-        ).filter(*base_filters)
-        ordered_query = page_query.order_by(
-            rated_at_expression.desc(),
-            LearnLessonFeedback.id.desc(),
-        )
-        if sort_by == "score_asc":
-            ordered_query = page_query.order_by(
-                LearnLessonFeedback.score.asc(),
-                rated_at_expression.desc(),
-                LearnLessonFeedback.id.desc(),
-            )
-        page_rows = ordered_query.offset(start).limit(safe_page_size).all()
-        user_map = _load_user_map(
-            sorted(
-                {
-                    str(getattr(row, "user_bid", "") or "").strip()
-                    for row in page_rows
-                    if str(getattr(row, "user_bid", "") or "").strip()
-                }
-            )
-        )
-
-        items: list[AdminOperationCourseRatingItemDTO] = []
-        for row in page_rows:
-            user_bid = str(getattr(row, "user_bid", "") or "").strip()
-            outline_item_bid = str(getattr(row, "outline_item_bid", "") or "").strip()
-            context = outline_context_map.get(
-                outline_item_bid,
-                {
-                    "chapter_outline_item_bid": "",
-                    "chapter_title": "",
-                    "lesson_outline_item_bid": outline_item_bid,
-                    "lesson_title": "",
-                },
-            )
-            user = user_map.get(user_bid, {})
-            items.append(
-                AdminOperationCourseRatingItemDTO(
-                    lesson_feedback_bid=str(
-                        getattr(row, "lesson_feedback_bid", "") or ""
-                    ),
-                    progress_record_bid=str(
-                        getattr(row, "progress_record_bid", "") or ""
-                    ),
-                    user_bid=user_bid,
-                    mobile=str(user.get("mobile", "") or ""),
-                    email=str(user.get("email", "") or ""),
-                    nickname=str(user.get("nickname", "") or ""),
-                    chapter_outline_item_bid=str(
-                        context.get("chapter_outline_item_bid", "") or ""
-                    ),
-                    chapter_title=str(context.get("chapter_title", "") or ""),
-                    lesson_outline_item_bid=str(
-                        context.get("lesson_outline_item_bid", "") or ""
-                    ),
-                    lesson_title=str(context.get("lesson_title", "") or ""),
-                    score=int(getattr(row, "score", 0) or 0),
-                    comment=str(getattr(row, "comment", "") or ""),
-                    mode=_resolve_course_rating_mode(
-                        str(getattr(row, "mode", "") or "")
-                    ),
-                    rated_at=_format_operator_datetime(getattr(row, "rated_at", None)),
-                )
-            )
-
-        if include_summary:
-            summary = AdminOperationCourseRatingSummaryDTO(
-                average_score=_format_average_score(
-                    getattr(summary_row, "average_score", None)
-                ),
-                rating_count=total,
-                user_count=int(getattr(summary_row, "user_count", 0) or 0),
-                latest_rated_at=_format_operator_datetime(
-                    getattr(summary_row, "latest_rated_at", None)
-                ),
-            )
-        else:
-            summary = AdminOperationCourseRatingSummaryDTO()
-        return AdminOperationCourseRatingListDTO(
-            summary=summary,
-            items=items,
-            page=safe_page_index,
-            page_size=safe_page_size,
-            total=total,
-            page_count=math.ceil(total / safe_page_size) if safe_page_size else 0,
-        )
-
-
-def get_operator_course_follow_up_detail(
-    app: Flask,
-    *,
-    shifu_bid: str,
-    generated_block_bid: str,
-) -> AdminOperationCourseFollowUpDetailDTO:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        normalized_generated_block_bid = str(generated_block_bid or "").strip()
-        if not normalized_shifu_bid:
-            raise_param_error("shifu_bid is required")
-        if not normalized_generated_block_bid:
-            raise_param_error("generated_block_bid is required")
-
-        detail_source, outline_items = _load_operator_course_outline_items(
-            normalized_shifu_bid
-        )
-        course = detail_source["course"]
-        outline_context_map = _build_course_outline_context_map(outline_items)
-        ask_block = (
-            LearnGeneratedBlock.query.filter(
-                LearnGeneratedBlock.shifu_bid == normalized_shifu_bid,
-                LearnGeneratedBlock.generated_block_bid
-                == normalized_generated_block_bid,
-                LearnGeneratedBlock.deleted == 0,
-                LearnGeneratedBlock.status == 1,
-                LearnGeneratedBlock.type == BLOCK_TYPE_MDASK_VALUE,
-                LearnGeneratedBlock.role == ROLE_STUDENT,
-            )
-            .order_by(LearnGeneratedBlock.id.desc())
-            .first()
-        )
-        if ask_block is None:
-            raise_param_error("generated_block_bid")
-
-        progress_record_bid = str(ask_block.progress_record_bid or "").strip()
-        groups = _load_follow_up_groups_for_progress_record(progress_record_bid)
-        selected_group_index = next(
-            (
-                index
-                for index, group in enumerate(groups)
-                if str(group["ask_block"].generated_block_bid or "").strip()
-                == normalized_generated_block_bid
-            ),
-            -1,
-        )
-        if selected_group_index < 0:
-            raise_param_error("generated_block_bid")
-
-        selected_group = groups[selected_group_index]
-        user_map = _load_user_map([str(ask_block.user_bid or "").strip()])
-        user = user_map.get(str(ask_block.user_bid or "").strip(), {})
-        context = outline_context_map.get(
-            str(ask_block.outline_item_bid or "").strip(),
-            {
-                "chapter_title": "",
-                "lesson_title": "",
-            },
-        )
-
-        timeline: list[AdminOperationCourseFollowUpTimelineItemDTO] = []
-        for index, group in enumerate(groups):
-            current_ask_block = group["ask_block"]
-            is_current = index == selected_group_index
-            timeline.append(
-                AdminOperationCourseFollowUpTimelineItemDTO(
-                    role="student",
-                    content=str(
-                        getattr(current_ask_block, "generated_content", "") or ""
-                    ),
-                    created_at=_format_operator_datetime(
-                        getattr(current_ask_block, "created_at", None)
-                    ),
-                    is_current=is_current,
-                )
-            )
-            answer_block = group.get("answer_block")
-            answer_content = _resolve_follow_up_answer_content(answer_block)
-            if answer_content:
-                timeline.append(
-                    AdminOperationCourseFollowUpTimelineItemDTO(
-                        role="teacher",
-                        content=answer_content,
-                        created_at=_format_operator_datetime(
-                            getattr(answer_block, "created_at", None)
-                        ),
-                        is_current=is_current,
-                    )
-                )
-
-        selected_answer_block = selected_group.get("answer_block")
-        source_info = _resolve_follow_up_source(
-            ask_block=ask_block,
-            answer_block=selected_answer_block,
-        )
-        return AdminOperationCourseFollowUpDetailDTO(
-            basic_info=AdminOperationCourseFollowUpDetailBasicInfoDTO(
-                generated_block_bid=normalized_generated_block_bid,
-                progress_record_bid=progress_record_bid,
-                user_bid=str(ask_block.user_bid or ""),
-                mobile=str(user.get("mobile", "") or ""),
-                email=str(user.get("email", "") or ""),
-                nickname=str(user.get("nickname", "") or ""),
-                course_name=str(getattr(course, "title", "") or ""),
-                shifu_bid=normalized_shifu_bid,
-                chapter_title=str(context.get("chapter_title", "") or ""),
-                lesson_title=str(context.get("lesson_title", "") or ""),
-                created_at=_format_operator_datetime(
-                    getattr(ask_block, "created_at", None)
-                ),
-                turn_index=selected_group_index + 1,
-            ),
-            current_record=AdminOperationCourseFollowUpCurrentRecordDTO(
-                follow_up_content=str(
-                    getattr(ask_block, "generated_content", "") or ""
-                ),
-                answer_content=_resolve_follow_up_answer_content(selected_answer_block),
-                source_output_content=str(
-                    source_info.get("source_output_content", "") or ""
-                ),
-                source_output_type=str(source_info.get("source_output_type", "") or ""),
-                source_position=int(source_info.get("source_position", 0) or 0),
-                source_element_bid=str(source_info.get("source_element_bid", "") or ""),
-                source_element_type=str(
-                    source_info.get("source_element_type", "") or ""
-                ),
-            ),
-            timeline=timeline,
-        )
-
-
-def get_operator_course_chapter_detail(
-    app: Flask,
-    *,
-    shifu_bid: str,
-    outline_item_bid: str,
-) -> AdminOperationCourseChapterDetailDTO:
-    with app.app_context():
-        normalized_shifu_bid = str(shifu_bid or "").strip()
-        normalized_outline_item_bid = str(outline_item_bid or "").strip()
-        if not normalized_shifu_bid:
-            raise_param_error("shifu_bid is required")
-        if not normalized_outline_item_bid:
-            raise_param_error("outline_item_bid is required")
-
-        detail_source, outline_items = _load_operator_course_outline_items(
-            normalized_shifu_bid
-        )
-        course = detail_source["course"]
-        outline_item_map = {
-            str(item.outline_item_bid or "").strip(): item
-            for item in outline_items
-            if str(item.outline_item_bid or "").strip()
-        }
-        outline_item = outline_item_map.get(normalized_outline_item_bid)
-        if outline_item is None:
-            raise_error("server.shifu.outlineItemNotFound")
-
-        llm_system_prompt, llm_system_prompt_source = _resolve_prompt_with_fallback(
-            outline_item=outline_item,
-            outline_item_map=outline_item_map,
-            course=course,
-            field_name="llm_system_prompt",
-        )
-        return AdminOperationCourseChapterDetailDTO(
-            outline_item_bid=normalized_outline_item_bid,
-            title=outline_item.title or "",
-            content=getattr(outline_item, "content", "") or "",
-            llm_system_prompt=llm_system_prompt,
-            llm_system_prompt_source=llm_system_prompt_source,
-        )
-
-
-def _load_bill_usage_record_map(
-    usage_bids: Sequence[str],
-) -> Dict[str, BillUsageRecord]:
-    normalized_usage_bids = sorted(
-        {
-            str(usage_bid or "").strip()
-            for usage_bid in usage_bids
-            if str(usage_bid or "").strip()
-        }
-    )
-    if not normalized_usage_bids:
-        return {}
-
-    rows = (
-        BillUsageRecord.query.filter(
-            BillUsageRecord.deleted == 0,
-            BillUsageRecord.usage_bid.in_(normalized_usage_bids),
-        )
-        .order_by(BillUsageRecord.id.desc())
-        .all()
-    )
-    usage_map: Dict[str, BillUsageRecord] = {}
-    for row in rows:
-        usage_bid = str(row.usage_bid or "").strip()
-        if usage_bid and usage_bid not in usage_map:
-            usage_map[usage_bid] = row
-    return usage_map
-
-
-def _build_latest_bill_usage_record_subquery(
-    *,
-    user_bid: str = "",
-    usage_bids: Sequence[str] | None = None,
-):
-    normalized_user_bid = str(user_bid or "").strip()
-    normalized_usage_bids = [
-        str(usage_bid or "").strip()
-        for usage_bid in usage_bids or []
-        if str(usage_bid or "").strip()
-    ]
-    query = db.session.query(
-        BillUsageRecord.usage_bid.label("usage_bid"),
-        db.func.max(BillUsageRecord.id).label("max_id"),
-    ).filter(
-        BillUsageRecord.deleted == 0,
-        BillUsageRecord.record_level == 0,
-    )
-    if normalized_user_bid:
-        query = query.filter(BillUsageRecord.user_bid == normalized_user_bid)
-    if usage_bids is not None and not normalized_usage_bids:
-        query = query.filter(false())
-    if normalized_usage_bids:
-        query = query.filter(BillUsageRecord.usage_bid.in_(normalized_usage_bids))
-    return query.group_by(BillUsageRecord.usage_bid).subquery()
-
-
-def _build_latest_billing_order_subquery(*, creator_bid: str):
-    normalized_creator_bid = str(creator_bid or "").strip()
-    return (
-        db.session.query(
-            BillingOrder.bill_order_bid.label("bill_order_bid"),
-            db.func.max(BillingOrder.id).label("max_id"),
-        )
-        .filter(
-            BillingOrder.deleted == 0,
-            BillingOrder.creator_bid == normalized_creator_bid,
-        )
-        .group_by(BillingOrder.bill_order_bid)
-        .subquery()
-    )
-
-
-def _find_operator_course_bids_by_name(course_name: str) -> Set[str]:
-    normalized_course_name = str(course_name or "").strip().lower()
-    if not normalized_course_name:
-        return set()
-
-    def _load_matching_bids(model) -> Set[str]:
-        latest_subquery = (
-            db.session.query(db.func.max(model.id).label("max_id"))
-            .filter(model.deleted == 0)
-            .group_by(model.shifu_bid)
-            .subquery()
-        )
-        rows = (
-            db.session.query(model.shifu_bid)
-            .join(latest_subquery, latest_subquery.c.max_id == model.id)
-            .filter(model.title.ilike(f"%{normalized_course_name}%"))
-            .all()
-        )
-        return {
-            str(shifu_bid or "").strip()
-            for (shifu_bid,) in rows
-            if str(shifu_bid or "").strip()
-        }
-
-    matching_bids: Set[str] = set()
-    matching_bids.update(_load_matching_bids(DraftShifu))
-    matching_bids.update(_load_matching_bids(PublishedShifu))
-    return matching_bids
-
-
-def _build_operator_course_query_filter(
-    shifu_bid_column: Any,
-    course_query: str,
-) -> Any | None:
-    normalized_course_query = str(course_query or "").strip()
-    if not normalized_course_query:
-        return None
-
-    course_filters = [shifu_bid_column == normalized_course_query]
-    matching_course_bids = _find_operator_course_bids_by_name(normalized_course_query)
-    if matching_course_bids:
-        course_filters.append(shifu_bid_column.in_(sorted(matching_course_bids)))
-    return or_(*course_filters)
-
-
-def _build_operator_course_overview(app: Flask) -> AdminOperationCourseOverviewDTO:
-    if not _can_use_operator_course_sql_optimization(app):
-        return _build_operator_course_overview_legacy(app)
-
-    candidate_query = _build_operator_course_candidate_query(
-        shifu_bid="",
-        course_name="",
-        creator_bids=None,
-        start_time=None,
-        end_time=None,
-    )
-    if candidate_query is None:
-        return AdminOperationCourseOverviewDTO()
-    candidate_subquery = candidate_query.subquery("operator_course_overview_candidates")
-    now = datetime.now()
-    created_window_start, created_window_end = _resolve_created_last_7d_window(now)
-    recent_activity_window_start = now - timedelta(days=30)
-    aggregate_row = db.session.query(
-        db.func.count(candidate_subquery.c.shifu_bid).label("total_course_count"),
-        db.func.sum(
-            case(
-                (candidate_subquery.c.course_status == COURSE_STATUS_UNPUBLISHED, 1),
-                else_=0,
-            )
-        ).label("draft_course_count"),
-        db.func.sum(
-            case(
-                (candidate_subquery.c.course_status == COURSE_STATUS_PUBLISHED, 1),
-                else_=0,
-            )
-        ).label("published_course_count"),
-        db.func.sum(
-            case(
-                (
-                    and_(
-                        candidate_subquery.c.created_at >= created_window_start,
-                        candidate_subquery.c.created_at <= created_window_end,
-                    ),
-                    1,
-                ),
-                else_=0,
-            )
-        ).label("created_last_7d_course_count"),
-    ).one()
-    total_course_count = int(aggregate_row.total_course_count or 0)
-    if total_course_count == 0:
-        return AdminOperationCourseOverviewDTO()
-    learning_active_30d_course_count = (
-        db.session.query(db.func.count(db.distinct(candidate_subquery.c.shifu_bid)))
-        .select_from(candidate_subquery)
-        .join(
-            LearnProgressRecord,
-            and_(
-                LearnProgressRecord.shifu_bid == candidate_subquery.c.shifu_bid,
-                LearnProgressRecord.deleted == 0,
-                LearnProgressRecord.status != LEARN_STATUS_RESET,
-                LearnProgressRecord.created_at >= recent_activity_window_start,
-            ),
-        )
-        .scalar()
-        or 0
-    )
-    paid_order_30d_course_count = (
-        db.session.query(db.func.count(db.distinct(candidate_subquery.c.shifu_bid)))
-        .select_from(candidate_subquery)
-        .join(
-            Order,
-            and_(
-                Order.shifu_bid == candidate_subquery.c.shifu_bid,
-                Order.deleted == 0,
-                Order.status == ORDER_STATUS_SUCCESS,
-                Order.created_at >= recent_activity_window_start,
-            ),
-        )
-        .scalar()
-        or 0
-    )
-
-    return AdminOperationCourseOverviewDTO(
-        total_course_count=total_course_count,
-        draft_course_count=int(aggregate_row.draft_course_count or 0),
-        published_course_count=int(aggregate_row.published_course_count or 0),
-        created_last_7d_course_count=int(
-            aggregate_row.created_last_7d_course_count or 0
-        ),
-        learning_active_30d_course_count=int(learning_active_30d_course_count or 0),
-        paid_order_30d_course_count=int(paid_order_30d_course_count or 0),
-    )
-
-
-def get_operator_course_overview(app: Flask) -> AdminOperationCourseOverviewDTO:
-    with app.app_context():
-        return _build_operator_course_overview(app)
-
-
-def _can_use_operator_course_sql_optimization(app: Flask) -> bool:
-    try:
-        return current_app._get_current_object() is app and db.engine is not None
-    except (RuntimeError, KeyError):
-        return False
-
-
-def _build_operator_course_overview_legacy(
-    app: Flask,
-) -> AdminOperationCourseOverviewDTO:
-    draft_rows = _load_latest_shifus(
-        DraftShifu,
-        shifu_bid="",
-        course_name="",
-        creator_bids=None,
-        start_time=None,
-        end_time=None,
-        updated_start_time=None,
-        updated_end_time=None,
-    )
-    published_rows = _load_latest_shifus(
-        PublishedShifu,
-        shifu_bid="",
-        course_name="",
-        creator_bids=None,
-        start_time=None,
-        end_time=None,
-        updated_start_time=None,
-        updated_end_time=None,
-    )
-    merged_courses, published_bids, _ = _merge_courses(draft_rows, published_rows)
-    total_course_count = len(merged_courses)
-    if total_course_count == 0:
-        return AdminOperationCourseOverviewDTO()
-
-    now = datetime.now()
-    created_window_start, created_window_end = _resolve_created_last_7d_window(now)
-    recent_activity_window_start = now - timedelta(days=30)
-    visible_shifu_bids = [
-        str(course.shifu_bid or "").strip()
-        for course in merged_courses
-        if str(course.shifu_bid or "").strip()
-    ]
-    learning_active_30d_course_count = len(
-        _load_recent_learning_active_course_bids(
-            since=recent_activity_window_start,
-            shifu_bids=visible_shifu_bids,
-        )
-    )
-    paid_order_30d_course_count = len(
-        _load_recent_paid_order_course_bids(
-            since=recent_activity_window_start,
-            shifu_bids=visible_shifu_bids,
-        )
-    )
-
-    return AdminOperationCourseOverviewDTO(
-        total_course_count=total_course_count,
-        draft_course_count=sum(
-            1
-            for course in merged_courses
-            if _resolve_course_status(course.shifu_bid or "", published_bids)
-            == COURSE_STATUS_UNPUBLISHED
-        ),
-        published_course_count=sum(
-            1
-            for course in merged_courses
-            if _resolve_course_status(course.shifu_bid or "", published_bids)
-            == COURSE_STATUS_PUBLISHED
-        ),
-        created_last_7d_course_count=sum(
-            1
-            for course in merged_courses
-            if course.created_at
-            and created_window_start <= course.created_at <= created_window_end
-        ),
-        learning_active_30d_course_count=learning_active_30d_course_count,
-        paid_order_30d_course_count=paid_order_30d_course_count,
-    )
-
-
-def _list_operator_courses_legacy(
-    app: Flask,
-    page_index: int,
-    page_size: int,
-    filters: Optional[dict] = None,
-) -> AdminOperationCourseListDTO:
-    safe_page_index = max(int(page_index or 1), 1)
-    safe_page_size = max(int(page_size or 20), 1)
-    filters = filters or {}
-
-    shifu_bid = str(filters.get("shifu_bid", "") or "").strip()
-    course_name = str(filters.get("course_name", "") or "").strip()
-    course_status = str(filters.get("course_status", "") or "").strip().lower()
-    quick_filter = _resolve_course_quick_filter(filters.get("quick_filter", ""))
-    creator_keyword = str(filters.get("creator_keyword", "") or "").strip()
-    start_time = filters.get("start_time")
-    end_time = filters.get("end_time")
-    updated_start_time = filters.get("updated_start_time")
-    updated_end_time = filters.get("updated_end_time")
-
-    creator_bids = _find_matching_creator_bids(creator_keyword)
-    draft_rows = _load_latest_shifu_seeds(
-        DraftShifu,
-        shifu_bid=shifu_bid,
-        course_name=course_name,
-        creator_bids=creator_bids,
-        start_time=start_time,
-        end_time=end_time,
-        updated_start_time=None,
-        updated_end_time=None,
-    )
-    published_rows = _load_latest_shifu_seeds(
-        PublishedShifu,
-        shifu_bid=shifu_bid,
-        course_name=course_name,
-        creator_bids=creator_bids,
-        start_time=start_time,
-        end_time=end_time,
-        updated_start_time=None,
-        updated_end_time=None,
-    )
-
-    merged_courses, published_bids, selected_sources = _merge_courses(
-        draft_rows, published_rows
-    )
-    activity_map = _load_course_activity_map(draft_rows, published_rows)
-
-    def resolve_activity(course) -> Dict[str, Any]:
-        return activity_map.get(str(course.shifu_bid or "").strip(), {})
-
-    def resolve_updated_at(course) -> Optional[datetime]:
-        activity = resolve_activity(course)
-        return activity.get("updated_at") or course.updated_at
-
-    if course_status in {COURSE_STATUS_PUBLISHED, COURSE_STATUS_UNPUBLISHED}:
-        merged_courses = [
-            course
-            for course in merged_courses
-            if _resolve_course_status(course.shifu_bid or "", published_bids)
-            == course_status
-        ]
-    if updated_start_time:
-        merged_courses = [
-            course
-            for course in merged_courses
-            if (resolve_updated_at(course) or datetime.min) >= updated_start_time
-        ]
-    if updated_end_time:
-        merged_courses = [
-            course
-            for course in merged_courses
-            if (resolve_updated_at(course) or datetime.min) <= updated_end_time
-        ]
-    if quick_filter:
-        if quick_filter == COURSE_QUICK_FILTER_DRAFT:
-            merged_courses = [
-                course
-                for course in merged_courses
-                if _resolve_course_status(course.shifu_bid or "", published_bids)
-                == COURSE_STATUS_UNPUBLISHED
-            ]
-        elif quick_filter == COURSE_QUICK_FILTER_PUBLISHED:
-            merged_courses = [
-                course
-                for course in merged_courses
-                if _resolve_course_status(course.shifu_bid or "", published_bids)
-                == COURSE_STATUS_PUBLISHED
-            ]
-        elif quick_filter == COURSE_QUICK_FILTER_CREATED_LAST_7D:
-            created_window_start, created_window_end = _resolve_created_last_7d_window()
-            merged_courses = [
-                course
-                for course in merged_courses
-                if course.created_at
-                and created_window_start <= course.created_at <= created_window_end
-            ]
-        else:
-            visible_shifu_bids = [
-                str(course.shifu_bid or "").strip()
-                for course in merged_courses
-                if str(course.shifu_bid or "").strip()
-            ]
-            if quick_filter == COURSE_QUICK_FILTER_LEARNING_ACTIVE_30D:
-                matched_shifu_bids = _load_recent_learning_active_course_bids(
-                    since=datetime.now() - timedelta(days=30),
-                    shifu_bids=visible_shifu_bids,
-                )
-            else:
-                matched_shifu_bids = _load_recent_paid_order_course_bids(
-                    since=datetime.now() - timedelta(days=30),
-                    shifu_bids=visible_shifu_bids,
-                )
-            merged_courses = [
-                course
-                for course in merged_courses
-                if str(course.shifu_bid or "").strip() in matched_shifu_bids
-            ]
-    merged_courses = sorted(
-        merged_courses,
-        key=lambda item: (
-            resolve_updated_at(item) or datetime.min,
-            item.created_at or datetime.min,
-            item.shifu_bid or "",
-        ),
-        reverse=True,
-    )
-    total = len(merged_courses)
-    page_offset = (safe_page_index - 1) * safe_page_size
-    page_items = merged_courses[page_offset : page_offset + safe_page_size]
-    draft_page_items = [
-        course
-        for course in page_items
-        if selected_sources.get(str(course.shifu_bid or "").strip()) == "draft"
-    ]
-    published_page_items = [
-        course
-        for course in page_items
-        if selected_sources.get(str(course.shifu_bid or "").strip()) == "published"
-    ]
-    _attach_course_prompt_flags(DraftShifu, draft_page_items)
-    _attach_course_prompt_flags(PublishedShifu, published_page_items)
-
-    user_bids = {
-        user_bid
-        for course in page_items
-        for user_bid in [
-            course.created_user_bid,
-            resolve_activity(course).get("updated_user_bid") or course.updated_user_bid,
-        ]
-        if user_bid and user_bid != "system"
-    }
-    user_map = _load_user_map(list(user_bids))
-    items = [
-        _build_course_summary(
-            course,
-            user_map,
-            _resolve_course_status(course.shifu_bid or "", published_bids),
-            resolve_activity(course),
-        )
-        for course in page_items
-    ]
-    return AdminOperationCourseListDTO(
-        items=items,
-        page=safe_page_index,
-        page_size=safe_page_size,
-        total=total,
-        page_count=((total + safe_page_size - 1) // safe_page_size) if total else 0,
-    )
-
-
-def list_operator_courses(
-    app: Flask,
-    page_index: int,
-    page_size: int,
-    filters: Optional[dict] = None,
-) -> AdminOperationCourseListDTO:
-    with app.app_context():
-        if not _can_use_operator_course_sql_optimization(app):
-            return _list_operator_courses_legacy(app, page_index, page_size, filters)
-
-        safe_page_index = max(int(page_index or 1), 1)
-        safe_page_size = max(int(page_size or 20), 1)
-        filters = filters or {}
-
-        shifu_bid = str(filters.get("shifu_bid", "") or "").strip()
-        course_name = str(filters.get("course_name", "") or "").strip()
-        course_status = str(filters.get("course_status", "") or "").strip().lower()
-        quick_filter = _resolve_course_quick_filter(filters.get("quick_filter", ""))
-        creator_keyword = str(filters.get("creator_keyword", "") or "").strip()
-        start_time = filters.get("start_time")
-        end_time = filters.get("end_time")
-        updated_start_time = filters.get("updated_start_time")
-        updated_end_time = filters.get("updated_end_time")
-
-        creator_bids = _find_matching_creator_bids(creator_keyword)
-        candidate_query = _build_operator_course_candidate_query(
-            shifu_bid=shifu_bid,
-            course_name=course_name,
-            creator_bids=creator_bids,
-            start_time=start_time,
-            end_time=end_time,
-            include_activity=True,
-        )
-        if candidate_query is None:
-            return AdminOperationCourseListDTO(
-                items=[],
-                page=safe_page_index,
-                page_size=safe_page_size,
-                total=0,
-                page_count=0,
-            )
-        candidate_subquery = candidate_query.subquery("operator_course_candidates")
-        query = db.session.query(candidate_subquery)
-
-        if course_status in {COURSE_STATUS_PUBLISHED, COURSE_STATUS_UNPUBLISHED}:
-            query = query.filter(candidate_subquery.c.course_status == course_status)
-        if quick_filter:
-            if quick_filter == COURSE_QUICK_FILTER_DRAFT:
-                query = query.filter(
-                    candidate_subquery.c.course_status == COURSE_STATUS_UNPUBLISHED
-                )
-            elif quick_filter == COURSE_QUICK_FILTER_PUBLISHED:
-                query = query.filter(
-                    candidate_subquery.c.course_status == COURSE_STATUS_PUBLISHED
-                )
-            elif quick_filter == COURSE_QUICK_FILTER_CREATED_LAST_7D:
-                created_window_start, created_window_end = (
-                    _resolve_created_last_7d_window()
-                )
-                query = query.filter(
-                    candidate_subquery.c.created_at >= created_window_start,
-                    candidate_subquery.c.created_at <= created_window_end,
-                )
-            else:
-                if quick_filter == COURSE_QUICK_FILTER_LEARNING_ACTIVE_30D:
-                    active_course_query = db.session.query(
-                        LearnProgressRecord.shifu_bid
-                    ).filter(
-                        LearnProgressRecord.deleted == 0,
-                        LearnProgressRecord.status != LEARN_STATUS_RESET,
-                        LearnProgressRecord.created_at
-                        >= datetime.now() - timedelta(days=30),
-                    )
-                    query = query.filter(
-                        candidate_subquery.c.shifu_bid.in_(active_course_query)
-                    )
-                else:
-                    paid_course_query = db.session.query(Order.shifu_bid).filter(
-                        Order.deleted == 0,
-                        Order.status == ORDER_STATUS_SUCCESS,
-                        Order.created_at >= datetime.now() - timedelta(days=30),
-                    )
-                    query = query.filter(
-                        candidate_subquery.c.shifu_bid.in_(paid_course_query)
-                    )
-
-        if updated_start_time:
-            query = query.filter(
-                candidate_subquery.c.activity_updated_at >= updated_start_time
-            )
-        if updated_end_time:
-            query = query.filter(
-                or_(
-                    candidate_subquery.c.activity_updated_at.is_(None),
-                    candidate_subquery.c.activity_updated_at <= updated_end_time,
-                )
-            )
-
-        total = int(query.count() or 0)
-        page_offset = (safe_page_index - 1) * safe_page_size
-        page_rows = (
-            query.order_by(
-                candidate_subquery.c.activity_updated_at.desc(),
-                candidate_subquery.c.created_at.desc(),
-                candidate_subquery.c.shifu_bid.desc(),
-            )
-            .offset(page_offset)
-            .limit(safe_page_size)
-            .all()
-        )
-        page_items = [_build_operator_course_list_candidate(row) for row in page_rows]
-
-        draft_page_items = [
-            course for course in page_items if course.selected_source == "draft"
-        ]
-        published_page_items = [
-            course for course in page_items if course.selected_source == "published"
-        ]
-        _attach_course_prompt_flags(DraftShifu, draft_page_items)
-        _attach_course_prompt_flags(PublishedShifu, published_page_items)
-
-        def resolve_activity(course) -> Dict[str, Any]:
-            return {
-                "updated_at": course.activity_updated_at or course.updated_at,
-                "updated_user_bid": course.activity_updated_user_bid
-                or course.updated_user_bid,
-            }
-
-        user_bids = {
-            user_bid
-            for course in page_items
-            for user_bid in [
-                course.created_user_bid,
-                resolve_activity(course).get("updated_user_bid")
-                or course.activity_updated_user_bid
-                or course.updated_user_bid,
-            ]
-            if user_bid and user_bid != "system"
-        }
-        user_map = _load_user_map(list(user_bids))
-        items = [
-            _build_course_summary(
-                course,
-                user_map,
-                course.course_status,
-                resolve_activity(course),
-            )
-            for course in page_items
-        ]
-        return AdminOperationCourseListDTO(
-            items=items,
-            page=safe_page_index,
-            page_size=safe_page_size,
-            total=total,
-            page_count=((total + safe_page_size - 1) // safe_page_size) if total else 0,
-        )
+from flaskr.service.shifu.admin_operations import courses as _operator_courses
+
+# Backward-compatible exports for existing imports from shifu.admin.
+from flaskr.service.shifu.admin_operations.courses import (
+    _resolve_operator_credit_grant_type, _format_decimal, _coerce_operator_datetime,
+    _format_operator_datetime, _format_average_score, _resolve_course_rating_mode,
+    _resolve_course_rating_sort_by, _resolve_course_credit_usage_mode,
+    _resolve_course_credit_usage_mode_filter, _resolve_course_credit_usage_scene,
+    _resolve_course_credit_usage_scene_filter, _build_course_credit_usage_scene_filter,
+    _resolve_course_credit_usage_view, _build_course_credit_usage_model_display,
+    _build_course_credit_usage_group_key, _build_operator_course_credit_usage_item,
+    _build_course_credit_usage_generation_name_expr, _build_course_credit_usage_ask_filter,
+    _build_course_credit_usage_learn_filter,
+    _build_operator_course_credit_usage_ledger_totals_subquery,
+    _build_operator_course_credit_usage_base_query,
+    _resolve_course_credit_usage_output_summary,
+    _load_course_credit_usage_output_summary_map,
+    _build_operator_course_credit_usage_detail_item, _apply_course_credit_usage_filters,
+    _build_course_credit_usage_covered_completed_user_subquery,
+    _build_operator_course_credit_metrics, _normalize_metadata_json, _normalize_identifier,
+    _load_course_user_contact_map, _load_user_map, _resolve_course_user_role,
+    _resolve_course_user_learning_status, _build_course_order_amount_expr,
+    _find_matching_creator_bids, _load_operator_user_last_login_map,
+    OperatorCourseListSeed, OperatorCourseListCandidate, _build_operator_course_list_seed,
+    _build_operator_course_list_candidate, _build_operator_visible_course_filter,
+    _build_latest_operator_course_rows_query, _build_latest_operator_course_rows_subquery,
+    _build_operator_course_candidate_query, _build_latest_outline_activity_subquery,
+    _build_operator_course_latest_activity_subquery, _build_latest_shifus_query,
+    _load_latest_shifus, _load_latest_shifu_seeds, _attach_course_prompt_flags,
+    _build_course_summary, _is_operator_visible_course, _resolve_course_status,
+    _resolve_course_quick_filter, _resolve_created_last_7d_window,
+    _load_course_activity_map, _load_latest_course_for_transfer,
+    _load_latest_active_draft_outlines, _build_course_copy_title,
+    _resolve_course_copy_title, _build_outline_history_tree,
+    _copy_course_variable_definitions, _run_course_copy_draft_risk_check,
+    _run_course_copy_outline_risk_check, _validate_operator_target_contact,
+    _prepare_operator_target_creator, _load_recent_learning_active_course_bids,
+    _load_recent_paid_order_course_bids, _clear_shifu_permission_cache,
+    _clear_shifu_creator_cache, _update_course_creator_bid,
+    transfer_operator_course_creator, copy_operator_course, _merge_courses,
+    _load_latest_course_versions, _load_operator_course_detail_source,
+    _load_latest_outline_items, _resolve_learning_permission, _resolve_content_status,
+    _resolve_outline_prompt_source, _resolve_prompt_with_fallback, _build_chapter_tree,
+    _load_outline_learning_stats, _load_operator_course_outline_items,
+    _resolve_visible_leaf_outline_bids, _build_course_outline_context_map,
+    _build_course_follow_up_base_subquery, _build_follow_up_user_keyword_filter,
+    _build_credit_usage_user_keyword_filter, _resolve_follow_up_matching_outline_bids,
+    _resolve_follow_up_answer_block, _resolve_follow_up_answer_content,
+    _load_follow_up_groups_for_progress_record, _resolve_follow_up_source_from_element,
+    _resolve_follow_up_source_from_blocks, _resolve_follow_up_source,
+    _load_course_related_user_bids, _load_course_user_paid_amount_map,
+    _load_course_user_last_learning_map, _load_course_user_joined_at_map,
+    _load_course_user_learned_lesson_count_map, get_operator_course_detail,
+    get_operator_course_prompt, get_operator_course_users,
+    get_operator_course_credit_usages, get_operator_course_credit_usage_details,
+    get_operator_course_follow_ups, get_operator_course_ratings,
+    get_operator_course_follow_up_detail, get_operator_course_chapter_detail,
+    _load_bill_usage_record_map, _build_latest_bill_usage_record_subquery,
+    _build_latest_billing_order_subquery, _find_operator_course_bids_by_name,
+    _build_operator_course_query_filter, _build_operator_course_overview,
+    get_operator_course_overview, _can_use_operator_course_sql_optimization,
+    _build_operator_course_overview_legacy, _list_operator_courses_legacy,
+    list_operator_courses,
+)
+
+
+class _AdminCompatibilityModule(type(sys)):
+    def __setattr__(self, name, value):
+        super().__setattr__(name, value)
+        if hasattr(_operator_courses, name):
+            setattr(_operator_courses, name, value)
+
+
+sys.modules[__name__].__class__ = _AdminCompatibilityModule
+
+
+__all__ = (
+    "get_course_visit_count_30d",
+    "UNIT_TYPE_VALUE_GUEST",
+    "UNIT_TYPE_VALUE_NORMAL",
+    "UNIT_TYPE_VALUE_TRIAL",
+    "check_text_with_risk_control",
+    "run_creator_granted_post_auth",
+    "_resolve_operator_credit_grant_type",
+    "_format_average_score",
+    "_resolve_course_rating_mode",
+    "_resolve_course_rating_sort_by",
+    "_resolve_course_credit_usage_view",
+    "_build_course_credit_usage_model_display",
+    "_build_course_credit_usage_group_key",
+    "_build_operator_course_credit_usage_item",
+    "_load_course_credit_usage_output_summary_map",
+    "_build_operator_course_credit_usage_detail_item",
+    "_apply_course_credit_usage_filters",
+    "_build_operator_course_credit_metrics",
+    "_load_course_user_contact_map",
+    "_load_user_map",
+    "_resolve_course_user_role",
+    "_resolve_course_user_learning_status",
+    "_find_matching_creator_bids",
+    "_build_operator_course_list_candidate",
+    "_build_operator_course_candidate_query",
+    "_load_latest_shifu_seeds",
+    "_build_course_summary",
+    "_resolve_course_quick_filter",
+    "_resolve_created_last_7d_window",
+    "_load_course_activity_map",
+    "_load_latest_course_for_transfer",
+    "_load_latest_active_draft_outlines",
+    "_resolve_course_copy_title",
+    "_build_outline_history_tree",
+    "_copy_course_variable_definitions",
+    "_run_course_copy_draft_risk_check",
+    "_run_course_copy_outline_risk_check",
+    "_validate_operator_target_contact",
+    "_prepare_operator_target_creator",
+    "_load_recent_learning_active_course_bids",
+    "_load_recent_paid_order_course_bids",
+    "_clear_shifu_permission_cache",
+    "_clear_shifu_creator_cache",
+    "_update_course_creator_bid",
+    "transfer_operator_course_creator",
+    "copy_operator_course",
+    "_load_latest_course_versions",
+    "_load_operator_course_detail_source",
+    "_resolve_learning_permission",
+    "_resolve_content_status",
+    "_resolve_outline_prompt_source",
+    "_resolve_prompt_with_fallback",
+    "_build_chapter_tree",
+    "_load_outline_learning_stats",
+    "_load_operator_course_outline_items",
+    "_resolve_visible_leaf_outline_bids",
+    "_build_course_follow_up_base_subquery",
+    "_build_follow_up_user_keyword_filter",
+    "_resolve_follow_up_matching_outline_bids",
+    "_resolve_follow_up_answer_block",
+    "_resolve_follow_up_answer_content",
+    "_load_follow_up_groups_for_progress_record",
+    "_resolve_follow_up_source_from_element",
+    "_resolve_follow_up_source_from_blocks",
+    "_resolve_follow_up_source",
+    "_load_course_related_user_bids",
+    "_load_course_user_paid_amount_map",
+    "_load_course_user_last_learning_map",
+    "_load_course_user_joined_at_map",
+    "_load_course_user_learned_lesson_count_map",
+    "get_operator_course_detail",
+    "get_operator_course_prompt",
+    "get_operator_course_users",
+    "get_operator_course_credit_usages",
+    "get_operator_course_credit_usage_details",
+    "get_operator_course_follow_ups",
+    "get_operator_course_ratings",
+    "get_operator_course_follow_up_detail",
+    "get_operator_course_chapter_detail",
+    "_load_bill_usage_record_map",
+    "_build_latest_billing_order_subquery",
+    "_find_operator_course_bids_by_name",
+    "_build_operator_course_query_filter",
+    "_build_operator_course_overview",
+    "get_operator_course_overview",
+    "_can_use_operator_course_sql_optimization",
+    "_build_operator_course_overview_legacy",
+    "_list_operator_courses_legacy",
+    "list_operator_courses",
+)
